### 2019-11-01 #1858 @danielskatz
👋 @benjaminbolling - You also submitted https://github.com/openjournals/joss-reviews/issues/1855 - can you explain why there are 2 submissions?
