### 2019-11-01 #1858 @benjaminbolling
Hi,

About the 2 versions - I uploaded one and got errors due to not having properly filled out affiliations. So I withdrew the first version (I thought), fixed the issue and re-uploaded.

/Benjamin

>
> ﻿
> 👋 @benjaminbolling - You also submitted #1855 - can you explain why there are 2 submissions?
>
