### 2019-11-01 #1858 @danielskatz
It looks like the references are not cited in the .md file, so while the paper compiles, there are no references found to be needed to be included in the paper.
