### 2019-11-01 #1858 @danielskatz
It is clear that this software is useful and is being used. The question for JOSS is whether it is being used in research. One additional question here is who you expect would cite this package/submission?
