### 2019-11-13 3753277
Updated paper.md to include references.
