### 2019-11-13 #1858 @benjaminbolling
Dear all,

first of all - I am sorry for taking an awful lot of time to respond but I have began working at a new position and was full with work to do.

I have now cited references in the .md file and slightly edited the text to do so.

The ones I would expect to cite this package or submission would be people working in control rooms at particle accelerators (perhaps other research or industrial facilities as well), including control room operators and physicists (such as accelerator physicists doing experiments with the accelerator itself and beamline scientists doing experiments on their side). The strength of this package is that it is open and dynamical. I intend to keep updating the package and will soon integrate the EPICS control system as well (after taking a course in EPICS programming in December).

Kind regards
