### 2019-11-14 #1858 @labarba
@benjaminbolling — I am the Associate Editor in Chief on rotation this week. I've looked at your paper and your explanation above, and I ask for your patience, because this submission is unusual for JOSS and may be difficult to put through review.

It sounds like this is more of an infrastructure tool, used in conjunction with large-scale physics equipment. Even if this is clearly within the scope of scientific workflows, it's not the kind of research software that JOSS commonly publishes. For that reason, we don't have an editor in this topic, and we may have trouble identifying reviewers.

Please help us by looking over the reviewer list, and telling us if you think you identify any potential reviewers. Also, give us some clues about how a reviewer might go about checking the functionality of the software, without access to the instrumentation it is meant to work with.
