### 2019-11-19 #1858 @benjaminbolling
I see, I will look through the reviewer list. The functionality of the software can be tested by selecting the so-called "Randomizer" control system, which I made for my own diagnostics/trials purposes when working from home and demonstration purposes as it only generates (not completely) random numerical values.
