### 2019-11-19 8c37faf
Mainly added update to check OS so that appropriate color schemes and proper text-to-speech system is used.
