+++ DynaGUI_Alarms.py
@@ -22,7 +22,7 @@ try:
     from PyQt5 import QtCore
 except:
     from PyQt4 import QtCore, QtGui
-import os
+import os, platform
 from datetime import datetime
 from functools import partial
 from time import sleep
@@ -47,31 +47,31 @@ class Dialog(QtGui.QDialog):
                             'r3-319l/dia/tco-02/temperature',
                             'r3-319l/dia/tco-03/temperature',
                             'r1-101/dia/bpm-01/xmeanpossa']
-            
+
             self.devdesc = ['R3 319 TCO 01 temperature',
                             'R3 319 TCO 02 temperature',
                             'R3 319 TCO 03 temperature',
                             'R1 101 BPM 01 x-pos']
-            
+
             self.devlims = [36,
                             38,
                             40,
                             100]
-            
+
             self.Nrows = 20
         self.reloadflag = 0
         self.maxsize = 0
         self.timerinterval = 30 # seconds
-        
+
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
         self.toplayout.addStretch()
-        
+
         # Construct a horizontal layout box for the edit and get all attribute buttons (must be a sublayer of the toplayout)
         self.editgetallwdg = QtGui.QWidget(self)
         self.toplayout.addWidget(self.editgetallwdg)
         self.horizlayout0 = QtGui.QHBoxLayout(self.editgetallwdg)
-        
+
         # Construct the button for setting up a dynamic list of attributes
         self.listbtn = QtGui.QPushButton("Edit DynaGUI")
         self.listbtn.clicked.connect(self.listbtnclicked)
@@ -81,30 +81,30 @@ class Dialog(QtGui.QDialog):
             self.doublevalidator = QtGui2.QDoubleValidator(-float('inf'),float('inf'),5)
         except:
             self.doublevalidator = QtGui.QDoubleValidator(-float('inf'),float('inf'),5)
-            
+
         # Now we construct the sublayout which will consist of the dynamically constructed buttons of the lists defined above (in example; list1 or list2)
         self.sublayout = QtGui.QGridLayout()
         self.toplayout.addLayout(self.sublayout)
-        
+
         # Now we construct a groupbox for all the dynamically constructed buttons. Edit its text to whatever is appropriate. Then its added to the sublayout.# Now we construct the sublayout which will consist of the dynamically constructed buttons of the lists defined above (in example; list1 or list2)
         self.groupBox = QtGui.QGroupBox()
         self.sublayout.addWidget(self.groupBox)
         self.sublayout = QtGui.QGridLayout(self.groupBox)
-        
+
         # Construct a simple label widget which in this example has the purpose of displaying various messages to the user (status messages)
         self.bottomlabel = QtGui.QLabel("")
         self.toplayout.addWidget(self.bottomlabel)
-        
+
         # Construct a horizontal layout box for the load and save buttons (must be a sublayer of the toplayout)
         self.loadsavewdg = QtGui.QWidget(self)
         self.toplayout.addWidget(self.loadsavewdg)
         self.horizlayout1 = QtGui.QHBoxLayout(self.loadsavewdg)
-        
+
         # Construct a horiztontal layout box for the Plot and Update buttons (must be a sublayer of the toplayout)
         self.plotupdwdg = QtGui.QWidget(self)
         self.toplayout.addWidget(self.plotupdwdg)
         self.horizlayout2 = QtGui.QHBoxLayout(self.plotupdwdg)
-        
+
         # Construct the load and save buttons, connect them to their functions and add them to their horizontal container
         self.loadbtn = QtGui.QPushButton("Load")
         self.savebtn = QtGui.QPushButton("Save")
[…]
