+++ DynaGUI_NV.py
@@ -22,15 +22,12 @@ try:
     from PyQt5 import QtCore
 except:
     from PyQt4 import QtCore, QtGui
-import os
+import os, platform, sys, time, datetime
 import numpy as np
 import pyqtgraph as pg
-import sys
 from math import *
 from functools import partial
 import matplotlib.pyplot as plt
-import time
-import datetime
 
 class Dialog(QtGui.QWidget):
     def __init__(self, inp, ctrl_library):
@@ -53,7 +50,7 @@ class Dialog(QtGui.QWidget):
                                         'State',
                                         'SRAM_MD_MODES',
                                         'SRAM_RMS']
-            
+
             # Some list of devices. List consists of some random stuff
             self.devlist = ['r3-319s2/dia/dcct-01',
                             'r1-101s/dia/dcct-01',
@@ -67,57 +64,57 @@ class Dialog(QtGui.QWidget):
         self.toSpecminutes = 15
         self.toSpecupdateFrequency = 2
         self.showallhideflag = False
-        
+
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
         self.toplayout.addStretch()
-        
+
         # Construct the combobox for the list of attributes
         self.listofbpmattributeslistbox = QtGui.QComboBox(self)
         self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
         self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
         self.toplayout.addWidget(self.listofbpmattributeslistbox)
-        
+
         # Construct a horizontal layout box for the edit and get all attribute buttons (must be a sublayer of the toplayout)
         self.editgetallwdg = QtGui.QWidget(self)
         self.toplayout.addWidget(self.editgetallwdg)
         self.horizlayout0 = QtGui.QHBoxLayout(self.editgetallwdg)
-        
+
         # Construct the button for setting up a dynamic list of attributes
         self.listbtn = QtGui.QPushButton("Edit DynaGUI")
         self.listbtn.clicked.connect(self.listbtnclicked)
         self.listbtn.setEnabled(True)
         self.horizlayout0.addWidget(self.listbtn)
-        
+
         # Construct the button for getting all attributes of all devices
         self.getAllAttsBtn = QtGui.QPushButton("Get all attributes")
         self.getAllAttsBtn.clicked.connect(self.getAllAttsClicked)
         self.getAllAttsBtn.setEnabled(True)
         self.horizlayout0.addWidget(self.getAllAttsBtn)
-        
+
         # Now we construct the sublayout which will consist of the dynamically constructed buttons of the lists defined above (in example; list1 or list2)
         self.sublayout = QtGui.QGridLayout()
         self.toplayout.addLayout(self.sublayout)
-        
+
         # Now we construct a groupbox for all the dynamically constructed buttons. Edit its text to whatever is appropriate. Then its added to the sublayout.# Now we construct the sublayout which will consist of the dynamically constructed buttons of the lists defined above (in example; list1 or list2)
         self.groupBox = QtGui.QGroupBox()
         self.sublayout.addWidget(self.groupBox)
         self.sublayout = QtGui.QGridLayout(self.groupBox)
-        
+
         # Construct a simple label widget which in this example has the purpose of displaying various messages to the user (status messages)
         self.bottomlabel = QtGui.QLabel("")
         self.toplayout.addWidget(self.bottomlabel)
-        
+
[…]

[…]
