+++ DynaGUI_NV.py
@@ -44,20 +44,34 @@ class Dialog(QtGui.QWidget):
             except:
                 loadflag = 0
         if loadflag == 0:
-            # All device attributes needed (basically which can be TRUE or False)
-            self.listofbpmattributes = ['lifetime',
-                                        'xpossa',
-                                        'State',
-                                        'SRAM_MD_MODES',
-                                        'SRAM_RMS']
-
-            # Some list of devices. List consists of some random stuff
-            self.devlist = ['r3-319s2/dia/dcct-01',
-                            'r1-101s/dia/dcct-01',
-                            'r1-108/dia/bpm-02',
-                            'r1-109/dia/bpm-01',
-                            'r3-a110811cab14/dia/bbbz-01',
-                            'b105a/dia/bsmon-01']
+            if self.ctrl_library == "Tango": # Some MAX IV Laboratory devices and attributes
+                self.devlist = ['r3-319s2/dia/dcct-01',
+                                'r1-101s/dia/dcct-01',
+                                'r1-108/dia/bpm-02',
+                                'r1-109/dia/bpm-01',
+                                'r3-a110811cab14/dia/bbbz-01',
+                                'b105a/dia/bsmon-01']
+                self.listofbpmattributes = ['lifetime',
+                                            'xpossa',
+                                            'State',
+                                            'SRAM_MD_MODES',
+                                            'SRAM_RMS']
+            elif self.ctrl_library == "EPICS": # Some ESS PV:s
+                self.PV_list = ['LEBT-010:PwrC-PSCH-01:CurR',
+                                'LEBT-010:PwrC-PSCV-01:CurR',
+                                'LEBT-010:PwrC-SolPS-01:CurR',
+                                'LEBT-010:PwrC-SolPS-02:CurR']
+                self.PV_descriptions = ['LEBT_CH-01_Current',
+                                        'LEBT_CV-01_Current',
+                                        'LEBT_Sol-01_Current',
+                                        'LEBT_Sol-02_Current']
+            elif ctrl_library == "Randomizer":
+                self.devlist = []
+                for m in range(0,35):
+                    self.devlist.append(str("Random_device_"+str(m)))
+                self.listofbpmattributes = []
+                for m in range(0,8):
+                    self.listofbpmattributes.append(str("Random_attribute_"+str(m)))
             self.Nrows = 15
         self.reloadflag = 0
         self.maxsize = 0
@@ -70,10 +84,17 @@ class Dialog(QtGui.QWidget):
         self.toplayout.addStretch()
 
         # Construct the combobox for the list of attributes
-        self.listofbpmattributeslistbox = QtGui.QComboBox(self)
-        self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
-        self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
-        self.toplayout.addWidget(self.listofbpmattributeslistbox)
+        if self.ctrl_library == "Tango":
+            self.listofbpmattributeslistbox = QtGui.QComboBox(self)
+            self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
+            self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
+            self.toplayout.addWidget(self.listofbpmattributeslistbox)
+        # elif self.ctrl_library == "EPICS": DO NOT add, EPICS uses PV:s and not devices with attributes
[…]
