### 2019-11-21 434df78
EPICS has been integrated within NV. The 1D plotting function also has EPICS working, and an EPICS-based archiving tool that can grab data from a DB and plot it. Bugs in the archiving mode exists and are currently being addressed.
