+++ DynaGUI_NV.py
@@ -22,7 +22,7 @@ try:
     from PyQt5 import QtCore
 except:
     from PyQt4 import QtCore, QtGui
-import os, platform, sys, time, datetime
+import os, platform, sys, time, datetime, fnmatch
 import numpy as np
 import pyqtgraph as pg
 from math import *
@@ -32,7 +32,7 @@ import matplotlib.pyplot as plt
 class Dialog(QtGui.QWidget):
     def __init__(self, inp, ctrl_library):
         QtGui.QDialog.__init__(self)
-        self.setWindowTitle("DynaGUI NV")
+        self.setWindowTitle("DynaGUI NV - "+str(ctrl_library))
         self.ctrl_library = ctrl_library
         if inp == 0:
             loadflag = 0
@@ -50,21 +50,25 @@ class Dialog(QtGui.QWidget):
                                 'r1-108/dia/bpm-02',
                                 'r1-109/dia/bpm-01',
                                 'r3-a110811cab14/dia/bbbz-01',
-                                'b105a/dia/bsmon-01']
+                                'b105a/dia/bsmon-01'
+                                ]
                 self.listofbpmattributes = ['lifetime',
                                             'xpossa',
                                             'State',
                                             'SRAM_MD_MODES',
-                                            'SRAM_RMS']
+                                            'SRAM_RMS'
+                                            ]
             elif self.ctrl_library == "EPICS": # Some ESS PV:s
                 self.PV_list = ['LEBT-010:PwrC-PSCH-01:CurR',
                                 'LEBT-010:PwrC-PSCV-01:CurR',
-                                'LEBT-010:PwrC-SolPS-01:CurR',
-                                'LEBT-010:PwrC-SolPS-02:CurR']
+                                #'LEBT-010:PwrC-SolPS-01:CurR',
+                                #'LEBT-010:PwrC-SolPS-02:CurR'
+                                ]
                 self.PV_descriptions = ['LEBT_CH-01_Current',
                                         'LEBT_CV-01_Current',
-                                        'LEBT_Sol-01_Current',
-                                        'LEBT_Sol-02_Current']
+                                        #'LEBT_Sol-01_Current',
+                                        #'LEBT_Sol-02_Current'
+                                        ]
             elif ctrl_library == "Randomizer":
                 self.devlist = []
                 for m in range(0,35):
@@ -248,7 +252,10 @@ class Dialog(QtGui.QWidget):
 
     def killdynamicbuttongroup(self):
         # Destroy / kill all buttons currently constructed in the buttongroup.
-        self.bottomlabel.setText(str("Loading " + str(self.listofbpmattributeslistbox.currentText()) + " statuses..."))
+        if self.ctrl_library == "EPICS":
+            self.bottomlabel.setText(str("Loading defined PV statuses..."))
+        else:
+            self.bottomlabel.setText(str("Loading " + str(self.listofbpmattributeslistbox.currentText()) + " statuses..."))
         for i in reversed(range(self.sublayout.count())):
             item = self.sublayout.itemAt(i)
             if isinstance(item, QtGui.QWidgetItem):
@@ -273,9 +280,11 @@ class Dialog(QtGui.QWidget):
         elif self.ctrl_library == "EPICS":
             alldevs = self.PV_list
 
-        for index in alldevs:
+        for n,index in enumerate(alldevs):
             rowcount += 1
-            button = QtGui.QPushButton(index, self.groupBox)
+            button = QtGui.QPushButton(str(self.PV_descriptions[n]), self.groupBox)
+            if self.ctrl_library == "EPICS":
+                button.setToolTip(index)
             textbox = QtGui.QLineEdit("-", self.groupBox)
             textbox.setEnabled(False)
[…]
