### 2019-11-22 c7d77b9
Added a time delayer to the NV 1D plotting function so that lines can be moved in either ways on the time (x) axis.
