### 2019-11-22 4af24b7
Minor bug fixes for the NV 1D time delayer and archiving mode.
