+++ DynaGUI_NV.py
@@ -1332,6 +1332,7 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
         self.data_y0 = []
         self.time_0 = -1
         self.equations = []
+        self.archivemode = 0
         if self.scalarflag == 0:
             for n, inp in enumerate(self.devslist):
                 if self.ctrl_library == "Tango":
@@ -1382,7 +1383,6 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
                 elif self.ctrl_library == "Randomizer":
                     val = random.random()
                 elif self.ctrl_library == "EPICS":
-                    #PV = epics.PV(str(self.devslist[n]), auto_monitor=True)
                     val = epics.PV(str(self.devslist[n]), auto_monitor=True).value
                 preindex = 0
                 for m in range(len(val)):
@@ -1459,10 +1459,12 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
             del self.data_x
             del self.data_y
             del self.data_y0
-            self.data_x = [0]
+            self.data_x = []
             self.data_y = []
+            self.data_y0 = []
             for n in range(self.cols):
                 self.curve[n].clear()
+                self.data_x.append([0])
                 self.data_y.append([0])
                 self.data_y0.append([0])
 
@@ -1483,7 +1485,8 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
             pass
 
     def acceptNewPlotSettings(self):
-        self.contWidget.plot.setXRange(-60 * 1.01 * self.graphTime,0)
+        if self.pyqtgraphtimer.isActive():
+            self.contWidget.plot.setXRange(-60 * 1.01 * self.graphTime,0)
         self.contWidget.plot.setLabel('left',self.ylabel,color='white',size = 30)
         reply = QtGui.QMessageBox.question(self, 'Equations application', 'Do you want the equations to be applied on previous values in the plot?', QtGui.QMessageBox.Yes,QtGui.QMessageBox.No)
         if reply == QtGui.QMessageBox.Yes:
@@ -1494,15 +1497,34 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
                     self.curve[n].setData(self.data_x[n], self.data_y[n], pen=pg.mkPen(self.colorind[n], width=2))
             elif self.scalarflag == 1:
                 for n in range(self.cols):
-                    if len(self.data_y[n]) > 0:
-                        for m in range(len(self.data_y[n])):
-                            self.data_y[n][m] = self.funccalculator(self.data_y0[n][m],self.equations[n],m)
-                        self.curve[n].setData(self.data_x[n], self.data_y[n], pen=pg.mkPen(self.colorind[n], width=2))
+                    for m in range(len(self.data_y[n])):
+                        self.data_y[n][m] = self.funccalculator(self.data_y0[n][m],self.equations[n],m)
+                    if n == 0:
+                        xmin = min(data_x_f[n])
+                        xmax = max(data_x_f[n])
+                    else:
+                        if min(data_x_f[n]) < xmin:
+                            xmin = min(data_x_f[n])
+                        if max(data_x_f[n]) > xmax:
+                            xmin = min(data_x_f[n])
+                    timestamps = []
+                    for timestamp in self.data_x[n]:
+                        timestamps.append(timestamp - time.time() + self.coldelays[n])
+                        self.curve[n].setData(timestamps,self.data_y[n], pen=pg.mkPen(self.colorind[n], width=2, style=QtCore.Qt.DotLine))
+                    #xtii = self.contWidget.plot.getAxis('bottom')
[…]
