### 2019-11-25 b97a672
Equations and delays can now be applied to realtime plotting data and archived plotted data. A bug was also eliminated preventing changing of line colors while in static [stopped or archived] mode.
