+++ DynaGUI_NV.py
@@ -180,7 +180,7 @@ class Dialog(QtGui.QWidget):
         self.getallDevs()
 
     def savebtnclicked(self):
-        nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File')
+        yesorno, nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File')
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
         else:
@@ -578,8 +578,11 @@ class Dialog(QtGui.QWidget):
             self.toSpecDevList = DevsNames
 
         if failflag == 0:
-            prep1D = prep1DGUI(self)
-            prep1D.exec_()
+            if self.archivingonly == 0:
+                prep1D = prep1DGUI(self)
+                prep1D.exec_()
+            else:
+                self.okflag = 1
             if self.okflag == 1:
                 lt = PyQtGraphPlotter(self)
                 lt.resize(800,600)
@@ -1144,6 +1147,7 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
         elif self.parent.archivingonly == 1:
             self.contWidget.plotbtn.setEnabled(False)
             self.contWidget.plotbtn.setText("Archiving Mode")
+            self.archivemode = 1
         self.contWidget.loadbtn.clicked.connect(self.loadclick)
         self.contWidget.savebtn.clicked.connect(self.saveclick)
         self.contWidget.showhidebtn.clicked.connect(self.PlotSettings)
@@ -1222,7 +1226,7 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
             btn.setStyleSheet("background-color:rgb("+col0+","+col1+","+col2+","+")");
             btn.setFixedWidth(20)
             chBox.setChecked(self.devsPlotting[ind])
-            chBox.stateChanged.connect(self.chBoxChange)
+            chBox.stateChanged.connect(self.chBoxCheck)
             btn.clicked.connect(partial(self.colorbtnRGBchange,str(ind)))
             try:
                 fm = QtGui2.QFontMetrics(chBox.font())
@@ -1256,6 +1260,9 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
         self.contWidget.plot.setLabel('left',self.ylabel,color='white',size = 30)
         self.scrollarea.setVisible(False)
 
+        if self.archivemode == 1:
+            self.loadclick()
+
     def showhidelegend(self):
         if self.contWidget.showhidelegends.text() == "Hide legend":
             self.contWidget.showhidelegends.setText("Show legend")
@@ -1287,19 +1294,31 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
                             col1 = str(cm[1])
                             col2 = str(cm[2])
                             btn.setStyleSheet("background-color:rgb("+col0+","+col1+","+col2+","+")");
+                    self.chBoxCheck()
+            else:
+                QtGui.QMessageBox.information(self,"Error",'Incorrect N of inputs')
 
-    def chBoxChange(self):
+    def chBoxCheck(self):
         devsPlotting = []
         for ind,chBox in enumerate(self.contWidget.chGroupBox.findChildren(QtGui.QCheckBox)):
             if chBox.isChecked() == 1:
                 devsPlotting.append(1)
-                self.curve[ind].setData(self.data_x[ind], self.data_y[ind])
+                timestamps = []
+                for timestamp in self.data_x[ind]:
+                    if self.archivemode == 0:
+                        timestamps.append(timestamp - time.time() + self.coldelays[ind])
+                    elif self.archivemode == 1:
+                        timestamps.append(timestamp + self.coldelays[ind])
[…]
