### 2019-11-26 0901f38
Added clock time for import of data through a DB / archiver.
