+++ DynaGUI_NV.py
@@ -24,6 +24,7 @@ except:
     from PyQt4 import QtCore, QtGui
 import os, platform, sys, time, datetime, fnmatch
 import numpy as np
+import numexpr
 import pyqtgraph as pg
 from math import *
 from functools import partial
@@ -576,7 +577,7 @@ class Dialog(QtGui.QWidget):
             self.archivingonly = 0
             self.toSpecTaurusList = TaurusList
             self.toSpecDevList = DevsNames
-
+        self.okflag = 0
         if failflag == 0:
             if self.archivingonly == 0:
                 prep1D = prep1DGUI(self)
@@ -624,8 +625,7 @@ class Dialog(QtGui.QWidget):
             self.killdynamicbuttongroup()
             self.getallDevs()
             # The layout should be minimal, so make it unrealistically small (x=10, y=10 [px]) and then resize to minimum.
-            self.setMaximumSize(10,10)
-            self.resize(self.sizeHint().width(), self.sizeHint().height())
+            self.resizeDynaGUI()
             self.reloadflag = 0
 
     def listbtnclicked(self):
@@ -656,14 +656,16 @@ class Dialog(QtGui.QWidget):
                     if n not in devlist:
                         devlist.append(n)
                 self.PV_list = devlist
-
             self.maxsize = 0
             self.killdynamicbuttongroup()
             self.getallDevs()
+            self.resizeDynaGUI()
+            self.reloadflag = 0
+
             # The layout should be minimal, so make it unrealistically small (x=10, y=10 [px]) and then resize to minimum.
+    def resizeDynaGUI(self):
             self.setMaximumSize(10,10)
             self.resize(self.sizeHint().width(), self.sizeHint().height())
-            self.reloadflag = 0
 
     def closeEvent(self, event):
         reply = QtGui.QMessageBox.question(self, 'Exit', 'Are you sure you want to exit? All unsaved data will be lost.', QtGui.QMessageBox.Yes,QtGui.QMessageBox.No)
@@ -730,6 +732,8 @@ class listbtnGUI(QtGui.QDialog):
         elif self.parent.ctrl_library == "Randomizer":
             if self.parent.devlist != self.newlistDevs:
                 self.parent.devlist = self.newlistDevs
+                self.parent.setMaximumSize(10,10)
+                self.parent.resize(self.sizeHint().width(), self.sizeHint().height())
                 self.parent.reloadflag = 1
 
             textAtts = str(self.textboxAttr.toPlainText())
@@ -1318,31 +1322,44 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
                 self.curve[ind].clear()
         self.devsPlotting = devsPlotting
 
-    def funccalculator(self,inputval,equation,nn,ind):
-        outputval = inputval
+    def funccalculator(self,linenum,equation,nn):
         if equation == 'none':
-            outputval = inputval
+            outputval = self.data_y0[linenum][nn]
         else:
-            try:
-                equationtrial2 = equation.split("AL_")
-                if equationtrial2[0] == '':
-                    del(equationtrial2[0])
-                    equationtrial3 = str("")
-                else:
-                    equationtrial3 = str(equationtrial2[0])
-                    del(equationtrial2[0])
-                for n in range(len(equationtrial2)):
-                    adding = equationtrial2[n].split("_LA")
-                    equationtrial3 = equationtrial3 + str(inputval[int(adding[0])][nn])
-                    if len(adding) > 1:
-                        equationtrial3 = equationtrial3 + str(adding[1])
-                equationtrial = equationtrial3.split('RV')
[…]
