### 2019-12-06 #1858 @majensen
@andrewfowlie - would you be willing to review this paper for [JOSS](https://joss.theoj.org)? I appreciated your consideration!
