### 2019-12-06 #1858 @majensen
@Chilipp, @alexpghayes - would either of you be able to review this paper for [JOSS](https://joss.theoj.org)? Please let me know on this thread -- thanks in advance!
