### 2019-12-06 #1858 @andrewfowlie
In general I'm very willing to review submissions to joss, but it's not my area of expertise either, so I cannot review it. I don't think I know anyone suitable so cannot offer an alternative either.
