### 2019-12-10 #1858 @majensen
Hey @pythonpanda2 - might you be able to review this one?
