### 2019-12-11 #1858 @majensen
@pythonpanda2  - much appreciated. My intuition (correct me if I'm wrong @benjaminbolling ) is this work addresses a need for a UI in a highly specialized field -- which makes it cool IMO and suitable for the journal-- and it's really the UI functionality and the packaging that would take up the bulk of the review. The package has a "dummy backend" (Randomizer - see [comment above](https://github.com/openjournals/joss-reviews/issues/1858#issuecomment-555366639)), which I think will really aid the review, as @benjaminbolling points out.

@benjaminbolling if you can scare up a colleague with knowledge of the backend instruments who would be willing to comment in the review, that would be great, but it's not a show stopper AFAIC at this stage.
