### 2019-12-11 #1942 @whedon
[REVIEW]: The DynaGUI package
**Submitting author:** @benjaminbolling (<a href="http://orcid.org/0000-0002-6650-5365">Benjamin Edward Bolling</a>)
**Repository:** <a href="https://github.com/benjaminbolling/DynaGUI" target ="_blank">https://github.com/benjaminbolling/DynaGUI</a>
**Version:** v1.0.1
