### 2019-12-16 benjaminbolling/DynaGUI#2 @pythonpanda2
Question on Contribution and authorship
@benjaminbolling

Seems like there are two main contributors to this repo. Why is the other contributor [@danielskatz] not listed in the authourship or in the acknowledgment ?
