### 2019-12-16 #1942 @pythonpanda2
https://github.com/benjaminbolling/DynaGUI/issues/2#issue-538141258
