### 2019-12-16 benjaminbolling/DynaGUI#2 @danielskatz
My only contribution was some copy-editing fixes in the paper as a JOSS Associate-Editor-in-Chief when this was submitted.  I should not be listed as a contributor to the paper or software.
