### 2019-12-24 benjaminbolling/DynaGUI#3 @pythonpanda2
Documentation for the repo 
@benjaminbolling

The three different component of the DynaGUI package has been explained the article. Lot of the information expressed in the paper has a visual component which is not easy to decipher just from Figure 1. I  went back to the repo and lot of the information is explained much more in detail in

https://github.com/benjaminbolling/DynaGUI/blob/master/DynaGUI_documentation.pdf

There is a disconnect. It would be nice to directly add a GitHub documentation carefully explaining things as simple as prerequisites (e.g. requirements.txt ), installation etc. and linking it up nicely with the paper to make the information accessible.
