### 2019-12-24 benjaminbolling/DynaGUI#4 @pythonpanda2
General query 
@benjaminbolling

It seems that the package has a strong dependency on  PyTango . How widely  is this package used in the domain addressed in this article ? Would you envision using the DynaGUI package in application where  PyTango packages  are not used ?
