### 2019-12-24 #1942 @pythonpanda2
https://github.com/benjaminbolling/DynaGUI/issues/4
