### 2020-01-23 #1942 @Chilipp
Dear @majensen and @benjaminbolling. My apologies for the delay and thank you very much for your patience! I am done with my review (so far) and note a high number of issues to work on concerning the DynaGUI package. As these are so many, I did not want to create an issue for each individual point, but rather want to give @benjaminbolling the chance to comment on each of these.

My general feeling is, that this repository is a collection of scripts that have a massive lack in documentation. There are no installation instructions, there is no quick start guide (not even a README in the repository), and there are only explanations about the GUI widgets, but not even about how to start the GUI. Please let me know @benjaminbolling and @majensen if you want me to open issues for the individual points listed below.

### General checks

> **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

I recommend to rename GNU_Licence.txt to LICENSE, as this is common for software packages. But this is not a blocking issue, of course.

> **Contribution and authorship:** Has the submitting author (@benjaminbolling) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I cannot judge this @majensen as the commit history is not present in the repository. Don't know how to handle this, but from the paper itself it seems like some others have been involved from [Tavares et. al](https://doi.org/10.1107/S1600577514011503)

### Functionality

> **Installation:** Does installation proceed as outlined in the documentation?

I could not find any installation instructions. Please create a README.md and provide some guidance for the user.

> **Functionality:** Have the functional claims of the software been confirmed?

I couldn't because I did not install it.

### Documentation

There exists a PDF-Document in the repository ([DynaGUI_documentation.pdf](https://github.com/benjaminbolling/DynaGUI/blob/master/DynaGUI_documentation.pdf)) which serves, I guess as the documentation. So I will judge based on this document.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

There is a System Requirements section, but no installation instructions.

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

No, there is no example usage. There are some screenshots of the different widgets, but I could not find any executable code in the documentation to verify the functionality of the package.

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

As there are only explanations of the various images, no, it is not explained well enough. There should be a piece of code somewhere stating how to generate and open the different widgets. How to establish a connection to the server, etc. The code also contains almost no documentation/comments.

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

I could not find any tests associated with the repository.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No, there are no community guidelines.

### Software paper

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

I could not find any comment on that in the paper.

> **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

The package also depends on matplotlib and numpy, which need to be cited.
