### 2020-01-23 #1942 @majensen
@benjaminbolling please remember also to review @pythonpanda2 's questions that were created as issues on the DynaGUI repo.  thanks!
