### 2020-02-11 #1942 @benjaminbolling
Hi,

I am rewriting some code and adding in-code explanations which can assist you during the reviewal. I am sorry for taking so much time, but I will let you know when it is ready together with a clear manual and installation routine such that the reviewal will go smoothly. I have also been busy with a lot of work, but I am working on this as much as I can.

Thanks for the patience!

> On 31 Jan 2020, at 15:38, Mark Jensen <notifications@github.com> wrote:
>
> Hi @benjaminbolling <https://github.com/benjaminbolling> - how are things proceeding?
>
