### 2020-05-24 #1942 @benjaminbolling
Det all,

Thank you for being so patient! I am ensuring there are no bugs in the code now, finishing some additional features of it and expecting it to be ready with instructions and a short article before the end of this month.

Again, thank you for the patience!

Kind regards
Benjamin

> On 20 May 2020, at 19:14, Mark Jensen <notifications@github.com> wrote:
>
> @benjaminbolling just want to check in here with you. Let me know how it's going.
>
