### 2020-05-31 73b15c2
Updated the DynaGUI package and its documentation.
