+++ paper.bib
@@ -21,3 +21,17 @@
     URL = {https://www.riverbankcomputing.com/software/pyqt/},
     year = 2016
 }
+
+@webpage{pyepicsReference,
+    title = {{PyEpics}: {I}nterface for the {C}hannel {A}ccess ({CA}) library of the {EPICS} {C}ontrol {S}ystem to the {P}ython {P}rogramming language,
+    URL = {https://cars9.uchicago.edu/software/python/pyepics3/},
+    year = 2014
+}
+
+@article{pandasReference,
+    title={pandas: a foundational Python library for data analysis and statistics},
+    author={Wes McKinney},
+    journal={Python for High Performance and Scientific Computing},
+    volume={14},
+    year={2011}
+}
