+++ paper.md
@@ -37,13 +37,13 @@ The DynAlarmsGUI (Dynamic Alarms GUI) has been designed to monitor numerical val
 
 DynaGUI NV is the most advance tool in the package, which for each device has two columns. The first column is a button with the server address of the device, whilst the value of the selected attribute is shown in the second column and hence enables users to inspect values in a simple and fast manner. DynaGUI NV also allows for launching a controller for the device (using AtkPanel for Tango Controls [@TangoCS]) or initializing 1D plots or 2D colormaps for the selected attribute for all devices for which the attribute is valid. Each device's control panel button becomes painted in lime-green if the attribute is valid, in fuchsia if the attribute is not existing for the device and maroon if a connection to the device cannot be established. The plot initialization automatically launches for all devices for which the attribute is valid. The plotting tool has been elaborately described in the DynaGUI documentation, with features included in the 1D graph tool being plotting read values in real time and also setting up and plotting equations (or functions) of the read values as well as combining the read values with one another. The 1D graph tool supports both scalar and vector (or waveform) plotting in real time, and plot data can be saved and loaded.
 
-The package aims to enable users to construct dynamic GUI:s for multiple purposes, and the author is open for implementing new functions and control systems on demand. For testing purposes, an artificial control system 'Randomizer' has been created displaying only random values. In this DynaGUI version, the DynaGUI panels are constructed using PyQt [@PyQtReference], with versions 4 and 5 supported.
+The package aims to enable users to construct dynamic GUI:s for multiple purposes, and the author is open for implementing new functions and control systems on demand. For testing purposes, an artificial control system 'Randomizer' has been created displaying only random values. In this DynaGUI version, the DynaGUI panels are constructed using PyQt [@PyQtReference], with versions 4 and 5 supported. The author is working on fully implementing the PyEPICS [@pyepicsReference] package as well as a new Finance package built on Pandas [@pandasReference] in order to have more sources to monitor live-stream data from and hence demonstrate the openness that serves as the core of the package.
 
 # Figures
 
-![A dynamic control panel of DynaGUI has been configured (left), from which a 1D realtime plot has been launched for the devices of which the HorProfile vector attribute is valid. This vector attribute reveals the horizontal profile of a particle beam.](figure1.png)
+![A dynamic control panel of DynaGUI NV has been configured (top-left), from which a 1D realtime plot has been launched for 4 artificial devices for a made-up attribute (right). Using this tool, two other lines have been set up as functions of two input data streams, with equations defined in the bottom-left figure and then plotted.](figure1.png)
 
 # Acknowledgements
-The author recognizes and wants to thank Jonas Petersson and Robin Svard, Accelerator Operators at the MAX IV Laboratory, for developing the original 2D Spectrogram Application for monitoring transverse beam position via Beam Position Monitors at the MAX IV Laboratory storage rings. The author used their codes as inspiration for developing a generic algorithm which can initialize a 2D spectrogram for any given list of devices with attributes. The author also wants to thank Bernhard Meirose, also an Accelerator Operator at the MAX IV Laboratory, for giving the inspiration and awakening the idea to creating this package.
+The author wants to thank Bernhard Meirose, Accelerator Operator at the MAX IV Laboratory, for the discussion that inspired and lead me into developing this package. The author also recognises and wants to thank Jonas Petersson and Robin Svard, Accelerator Operators at the MAX IV Laboratory, for developing the original 2D Spectrogram Application for monitoring transverse beam position via Beam Position Monitors at the MAX IV Laboratory storage rings.
 
 # References
