+++ dynagui-files/DynaGUI_NV.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """
-<A Dynamic Graphical User Interface package, which gives users a method to construct temporary, permanent and/or a set of GUI:s for users in a simple and fast manner combined with diagnostics tools (with advance 1D and 2D plotting methods).>
-    Copyright (C) <2019>  <Benjamin Edward Bolling>
+<A Dynamic Graphical User Interface package, which gives users a method to construct temporary, permanent and/or a set of GUI:s for users in a simple and fast manner combined with diagnostics tools with advance 1D and 2D plotting methods.>
+    Copyright (C) <2019-2020>  <Benjamin Edward Bolling>  <benjaminbolling@icloud.com>
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -22,13 +22,6 @@ try:
     from PyQt5 import QtCore
 except:
     from PyQt4 import QtCore, QtGui
-import os, platform, sys, time, datetime, fnmatch
-import numpy as np
-import numexpr
-import pyqtgraph as pg
-from math import *
-from functools import partial
-import matplotlib.pyplot as plt
 
 class Dialog(QtGui.QWidget):
     def __init__(self, inp, ctrl_library):
@@ -60,16 +53,25 @@ class Dialog(QtGui.QWidget):
                                             'SRAM_RMS'
                                             ]
             elif self.ctrl_library == "EPICS": # Some ESS PV:s
-                self.PV_list = ['LEBT-010:PwrC-PSCH-01:CurR',
-                                'LEBT-010:PwrC-PSCV-01:CurR',
+                self.PV_list = ['LAB:AMC-002-4:I1-YP-TR1-ArrayData',
+                                #'LAB:AMC-002-4:I1-YP-AvgValue',
+                                #'LAB:AMC-002-4:I2-YP-AvgValue',
                                 #'LEBT-010:PwrC-SolPS-01:CurR',
                                 #'LEBT-010:PwrC-SolPS-02:CurR'
                                 ]
-                self.PV_descriptions = ['LEBT_CH-01_Current',
-                                        'LEBT_CV-01_Current',
+                self.PV_descriptions = ['LAB wave 1',
+                                        #'LAB scalar 1',
+                                        #'LAB scalar 2',
                                         #'LEBT_Sol-01_Current',
                                         #'LEBT_Sol-02_Current'
                                         ]
+            elif ctrl_library == "Finance":
+                self.stocknames = ['AAPL',
+                                   'TSLA',
+                                   'FB',
+                                   'BAX',
+                                   'AVGO',
+                                   'UBER']
             elif ctrl_library == "Randomizer":
                 self.devlist = []
                 for m in range(0,35):
@@ -77,6 +79,15 @@ class Dialog(QtGui.QWidget):
                 self.listofbpmattributes = []
                 for m in range(0,8):
                     self.listofbpmattributes.append(str("Random_attribute_"+str(m)))
+            elif ctrl_library == "HistoricalData":
+                self.devlist = ["Define",
+                                "your",
+                                "Devices'",
+                                "Names",
+                                "Here",
+                                "Separated",
+                                "by",
+                                "Rows"]
             self.Nrows = 15
         self.reloadflag = 0
         self.maxsize = 0
@@ -94,12 +105,14 @@ class Dialog(QtGui.QWidget):
             self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
             self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
[…]
