+++ paper.bib
@@ -23,14 +23,14 @@
 }
 
 @webpage{pyepicsReference,
-    title = {{PyEpics}: {I}nterface for the {C}hannel {A}ccess ({CA}) library of the {EPICS} {C}ontrol {S}ystem to the {P}ython {P}rogramming language,
+    title = {{PyEpics}: {I}nterface for the {C}hannel {A}ccess ({CA}) library of the {EPICS} {C}ontrol {S}ystem to the {P}ython {P}rogramming language},
     URL = {https://cars9.uchicago.edu/software/python/pyepics3/},
     year = 2014
 }
 
 @article{pandasReference,
-    title={pandas: a foundational Python library for data analysis and statistics},
-    author={Wes McKinney},
+    title={p}andas: a foundational Python library for data analysis and statistics,
+    author=McKinney, W.,
     journal={Python for High Performance and Scientific Computing},
     volume={14},
     year={2011}
