+++ paper.bib
@@ -29,8 +29,8 @@
 }
 
 @article{pandasReference,
-    title={p}andas: a foundational Python library for data analysis and statistics,
-    author=McKinney, W.,
+    title={pandas: a foundational Python library for data analysis and statistics},
+    author={McKinney, W.},
     journal={Python for High Performance and Scientific Computing},
     volume={14},
     year={2011}
