+++ paper.md
@@ -35,7 +35,7 @@ DynaGUI TF (True/False) is dynamical in the sense that the user can insert the n
 
 The DynAlarmsGUI (Dynamic Alarms GUI) has been designed to monitor numerical values and notify user(s) if a condition is not fulfilled. To edit the alarms GUI, the user has to press Edit DynaGUI to open an edit-panel. In the left window of the edit-panel, the user has to define the list of devices' server domains and signals to monitor. In the right window of the edit-panel, the user has to define descriptions of the different alarm signals as they should appear in the DynAlarmsGUI control panel, and also the sweep-time has to be defined (how frequent the system should check the values). The DynALarmsGUI's dynamic panel can be divided into three columns for each device and attribute it monitors. The first column contains the descriptions of the alarms, the second column contains the numerical values from the last sweep, the fourth column contains the alarm limit, and the third column contains a combobox with two gaps (larger than or smaller than) which points in the way it should be between the read value and the alarm limit. All gap criterias that are fulfilled results in that the background-color of the alarm description is painted lime-green. If a gap criteria is not fulfilled, an alarm will go off by using the computer's speakers, showing a message in the DynAlarmsGUI's status bar and the alarm description's background color becomes red.
 
-DynaGUI NV is the most advance tool in the package, which for each device has two columns. The first column is a button with the server address of the device, whilst the value of the selected attribute is shown in the second column and hence enables users to inspect values in a simple and fast manner. DynaGUI NV also allows for launching a controller for the device (using AtkPanel for Tango Controls [@TangoCS]) or initializing 1D plots or 2D colormaps for the selected attribute for all devices for which the attribute is valid. Each device's control panel button becomes painted in lime-green if the attribute is valid, in fuchsia if the attribute is not existing for the device and maroon if a connection to the device cannot be established. The plot initialization automatically launches for all devices for which the attribute is valid. The plotting tool has been elaborately described in the DynaGUI documentation, with features included in the 1D graph tool being plotting read values in real time and also setting up and plotting equations (or functions) of the read values as well as combining the read values with one another. The 1D graph tool supports both scalar and vector (or waveform) plotting in real time, and plot data can be saved and loaded.
+DynaGUI NV is the most advance tool in the package, which for each device has two columns. The first column is a button with the server address of the device, whilst the value of the selected attribute is shown in the second column and hence enables users to inspect values in a simple and fast manner. DynaGUI NV also allows for launching a controller for the device (using AtkPanel for Tango Controls [@TangoCS]) or initializing 1D plots or 2D colormaps for the selected attribute for all devices for which the attribute is valid. Each device's control panel button becomes painted in lime-green if the attribute is valid, in fuchsia if the attribute is not existing for the device and maroon if a connection to the device cannot be established. The plot initialization automatically launches for all devices for which the attribute is valid. The plotting tool has been elaborately described in the DynaGUI documentation, with features included in the 1D graph tool being plotting read values in real time and also setting up and plotting equations (or functions) of the read values as well as combining the read values with one another, with examples of such attached in Figure 1. The 1D graph tool supports both scalar and vector (or waveform) plotting in real time, and plot data can be saved and loaded.
 
 The package aims to enable users to construct dynamic GUI:s for multiple purposes, and the author is open for implementing new functions and control systems on demand. For testing purposes, an artificial control system 'Randomizer' has been created displaying only random values. In this DynaGUI version, the DynaGUI panels are constructed using PyQt [@PyQtReference], with versions 4 and 5 supported. The author is working on fully implementing the PyEPICS [@pyepicsReference] package as well as a new Finance package built on Pandas [@pandasReference] in order to have more sources to monitor live-stream data from and hence demonstrate the openness that serves as the core of the package.
 
