### 2020-05-31 #1942 @benjaminbolling
Dear all,
@majensen @arfon @Chilipp
thanks for the enormous patience. I have now updated the package, attached a README file with installation instructions and how to launch it, a documentation on how to use the applications in the package, and also updated the article itself.
