### 2020-05-31 e62274a
Updated the article md file - typo
