### 2020-06-17 #1942 @majensen
Hi @pythonpanda2 - I hope the new job is going well. Do you have a sense of when you might have the opportunity to look at @benjaminbolling 's work? Thanks very much  -- @majensen
