### 2020-06-23 benjaminbolling/DynaGUI#3 @pythonpanda2
@benjaminbolling

When i did mention about specifying requirement, i should have been much more clear.  Please see the example below.

https://github.com/aspuru-guzik-group/chemical_vae/blob/master/requirements.txt

Then you can easily guide potential user with a small note on the README to use either conda or pip to install the per-requisite packages with out any ambiguity.
