### 2020-06-23 #1942 @pythonpanda2
https://github.com/benjaminbolling/DynaGUI/issues/3#issuecomment-647886871
