### 2020-06-25 #1942 @pythonpanda2
There are already some very well rounded points raised by @Chilipp in the last round  . I would wait to see if some of those points have been addresses to their satisfaction  (To avoid repetitions) and   i will stick to the one issue I have raised.

In addition, the funding sources should be duly acknowledged.
