### 2020-07-02 benjaminbolling/DynaGUI#5 @Chilipp
Fix formatting of README.md
This issue is part of the review in https://github.com/openjournals/joss-reviews/issues/1942

The README.md file claims to be markdown formatted (ending with `.md`) but it is definitely not as you cannot really read it easily on https://github.com/benjaminbolling/DynaGUI. This is not a blocking issue as the format of the installation instructions, etc. is not part of the JOSS requirements, but I strongly recommend to fix this. If you have strong feelings about the format of this file @benjaminbolling, you should at least rename it to `README.txt`.

Here is some guide how to make a proper README file: https://github.com/18F/open-source-guide/blob/18f-pages/pages/making-readmes-readable.md
