### 2020-07-02 benjaminbolling/DynaGUI#7 @Chilipp
Where is the Example usage?
Hi @benjaminbolling,

as part of the review in https://github.com/openjournals/joss-reviews/issues/1942, I could not find the section about the example usage.

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Could you please point me to the place where to find it?
