### 2020-07-02 benjaminbolling/DynaGUI#8 @Chilipp
Add API documentation
As a requirement for the review in https://github.com/openjournals/joss-reviews/issues/1942, please add an API documentation to your python source code.

See https://realpython.com/documenting-python-code/ for instance about what I mean with this. At least the central objects of your software should be documented properly.
