### 2020-07-02 benjaminbolling/DynaGUI#9 @Chilipp
Create community guidelines
> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

As a requirement for https://github.com/openjournals/joss-reviews/issues/1942, please add a CONTRIBUTING file, or something comparable

see https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors for some help
