### 2020-07-02 #1942 @Chilipp
Hi @benjaminbolling,

my apologies for the delay and thank you very much @benjaminbolling for your ongoing work on this!

I have the feeling that most of my points [from my comment above](https://github.com/openjournals/joss-reviews/issues/1942#issuecomment-577582216) have not been addressed (or at least I could not find them). As it takes me quite a while to go through everything, please address the individual issues in the future that I reference down below, or cite my comments and answer within this thread.
I apologize if I just did not find the information on some of the issues mentioned below. If so, please point me individually to the correct place where this is answered from your perspective or ask if there is need for clarification.

> **Contribution and authorship:**

- [ ] @majensen, as stated above, I cannot judge this as I don't see the commit history. So this is something you have to decide

> **Installation:** Does installation proceed as outlined in the documentation?

- [ ] As already pointed out by @pythonpanda2 in benjaminbolling/DynaGUI#3, you should to add the requirements to the installation instruction, etc., see also benjaminbolling/DynaGUI#6
- [ ] You should rewrite the README.md file to make it properly markdown-formatted benjaminbolling/DynaGUI#5

> **Functionality:** Have the functional claims of the software been confirmed?

- [ ] I was able to start it using the environment file from benjaminbolling/DynaGUI#6, but as there is no example usage, I could not really test it (see benjaminbolling/DynaGUI#7)

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

- [ ] I did add an environments.yml file in benjaminbolling/DynaGUI#6, but we would still need to add proper installation instructions to the README file (or somewhere else)

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

- [ ] Could you point me to the place where this has been added or where I can find something about it? see benjaminbolling/DynaGUI#7

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

- [ ] There is no API documentation see benjaminbolling/DynaGUI#8

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

- [ ] I could not find any tests associated with the repository.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

- [ ] No, there are still no community guidelines, see benjaminbolling/DynaGUI#9

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

- [ ] @benjaminbolling, could you please point me to the place where this is shown as I could not find it

> **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

- [ ] you are still not citing matplotlib and numpy in your paper. see https://www.scipy.org/citing.html#numpy and https://matplotlib.org/3.1.1/citing.html
