### 2020-07-07 #1942 @benjaminbolling
Thank you @Chilipp, I will be addressing these issues in the coming two weeks in the evenings. @majensen Thus, we can set a deadline for next Friday, July 17th.
