### 2020-07-14 a37d6b4
Added Finance 2D plotting functionality. Added control panel for Finance package and EPICS package (empty for now for EPICS). Added the EPICS and Finance packages to the alarms GUI.
