+++ Launcher.py
@@ -210,64 +210,64 @@ class Dialog(QtGui.QWidget):
     def viewdataNVclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_NV.py 'HistoricalData' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_NV.py 'HistoricalData' "+fp+"&")
         self.close()
         print("View data NV clicked")
     def epicsNVclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_NV.py 'EPICS' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_NV.py 'EPICS' "+fp+"&")
         self.close()
         print("EPICS NV clicked")
     def epicsTFclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_TF.py 'EPICS' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_TF.py 'EPICS' "+fp+"&")
         print("EPICS TF clicked")
     def epicsAlarmsclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_Alarms.py 'EPICS' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_Alarms.py 'EPICS' "+fp+"&")
         print("EPICS Alarms clicked")
     def tangoNVclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_NV.py 'Tango' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_NV.py 'Tango' "+fp+"&")
         print("Tango NV clicked")
     def tangoTFclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_TF.py 'Tango' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_TF.py 'Tango' "+fp+"&")
         print("Tango TF clicked")
     def tangoAlarmsclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_Alarms.py 'Tango' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_Alarms.py 'Tango' "+fp+"&")
         print("Tango Alarms clicked")
     def financeNVclicked(self):
         dirpath = os.path.abspath(os.path.dirname(__file__))
         fp = str(self.conffilepath.text())
-        os.system("python "+dirpath+"/dynagui-files/DynaGUI_NV.py 'Finance' "+fp+"&")
+        os.system("python3.7 "+dirpath+"/dynagui-files/DynaGUI_NV.py 'Finance' "+fp+"&")
         print("Finance NV clicked")
[…]

[…]
