### 2020-07-15 #1942 @benjaminbolling
@Chilipp : I have updated the code a bit to address issue#4: To decrease the dependency of PyTango. All codes now enables user to select between the packages "Tango" (using PyTango), "EPICS" (using PyEPICS), "Finance" (using Pandas Datareader) and "Randomizer" (for testing and demonstration purposes using random values). Other packages used for obtaining values from sources can be added in later stages with ease. I will address each issue in their own reference.

As for "Contribution and authorship", I did all of the code except for the original code for the "2D Spectrogram" part which I referenced in both the paper.md and the documentation file under Acknowledgements. If this is incorrect or if you need me to change something there, let me know.

Again, thanks for the patience!
