### 2020-07-15 benjaminbolling/DynaGUI#4 @benjaminbolling
In order to decrease the dependency of PyTango. All codes now enables user to select between the packages "Tango" (using PyTango), "EPICS" (using PyEPICS), "Finance" (using Pandas Datareader) and "Randomizer" (for testing and demonstration purposes using random values). Other packages used for obtaining values from various data polling sources can be added in later stages with ease. @pythonpanda2 : Since you opened this issue, I leave it up to you to decide if this issue is to be considered as closed or not.
