### 2020-07-15 benjaminbolling/DynaGUI#4 @benjaminbolling
As for how much PyTango is used: "The TANGO control system is a free open source device-oriented controls toolkit for controlling any kind of hardware or software and building SCADA systems" (from TANGO wikipedia page), and PyTango is a package used in Python for communicating with the TANGO control system.

EPICS (Experimental Physics and Industrial Control System) is used at some other facilities (one of which I am working at), and hence, pyEPICS is used in the Python package to communicate with this control system.

In order to demonstrate the openness of the DynaGUI package, the two additional 'data analysis' packages "Finance" and "Random" were added.
