### 2020-07-15 c527002
Added the requirements.txt file which includes all required and recommended Python packages.
