+++ dynagui-files/Cassandra_ImportData.py
@@ -23,41 +23,34 @@ import fnmatch, pytz
 
 class CassImp:
     def __init__(self):
-        
+
         self.datatimestamps = 0
         self.readvals = 0
-        
+        self.hdb_cluster = ['hdb-server-addr-0',
+                            'hdb-server-addr-1',
+                            'hdb-server-addr-2']
+        self.hdb_name = "hdb-server-csdb-0"
+
     def attr_wildcard(self,wildcard):
         query = "SELECT cs_name, domain, family, member, name FROM att_names"
         attrs = []
-        """
-        hdb_cluster_old = ['b-v-db-cn-0',
-                       'b-v-db-cn-1',
-                       'b-v-db-cn-2',
-                       'b-v-db-cn-3',
-                       'b-v-db-cn-4',
-                       'b-v-db-cn-5']
-        """
-        hdb_cluster = ['b-picard13-cas-0',
-                       'b-kirk13-cas-3',
-                       'b-kirk13-cas-4',
-                       'b-kirk13-cas-5']
-        cluster = Cluster(hdb_cluster)
+
+        cluster = Cluster(self.hdb_cluster)
         session = cluster.connect("hdb")
-        
+
         for row in self.execute_cql_querry(session, query):
-            if "g-v-csdb-0" in row.cs_name:
+            if self.hdb_name in row.cs_name:
                 attr = "/".join((row.domain, row.family, row.member, row.name))
                 attrs.append(attr)
         matches = [att for att in attrs if fnmatch.fnmatch(att, wildcard)]
         return matches
-        
+
     def execute_cql_querry(self, session, cql_querry):
         prepared_request = session.prepare(cql_querry)
         gen = session.execute(prepared_request)
         cql_rows = [row for row in gen]
         return cql_rows
-        
+
     def get_att_data_type(self, att_conf_id, session):
         """ Return HDB++ data_type from an att_conf_id """
         cql_querry = "SELECT data_type FROM att_conf WHERE att_conf_id = {}"
@@ -66,7 +59,7 @@ class CassImp:
         reply = self.execute_cql_querry(session, cql_querry)
         data_type = reply[0].data_type
         return data_type
-        
+
     def get_att_conf_id(self, attr_name, session):
         """ Return the HDB++ att_conf_id for one attribute """
         cql_querry = "SELECT att_conf_id FROM att_conf WHERE att_name= '{}'"
@@ -81,19 +74,15 @@ class CassImp:
             raise NameError(msg)
         att_conf_id = reply[0].att_conf_id
         return att_conf_id
-        
+
     def get_data(self, data_type, att_id, period):
             query = "SELECT * from att_{} WHERE att_conf_id = {} and period = '{}'"
             select_period = query.format(data_type, att_id, period)
-            hdb_cluster = ['b-picard13-cas-0',
-                       'b-kirk13-cas-3',
-                       'b-kirk13-cas-4',
-                       'b-kirk13-cas-5']
-            cluster = Cluster(hdb_cluster)
+            cluster = Cluster(self.hdb_cluster)
             session = cluster.connect("hdb")
             data = self.execute_cql_querry(session, select_period)
             return data
-    
+
     def readingdata(self,inp1,inp2,inp3):
         attr_name = inp1
         period_start =  datetime.strptime(inp2, "%Y-%m-%d")
@@ -101,7 +90,7 @@ class CassImp:
             period_end = datetime.now()
         else:
             period_end =  datetime.strptime(inp3, "%Y-%m-%d")
-        
+
         periods = [(period_start + timedelta(days=d)).strftime("%Y-%m-%d")
                    for d in range((period_end - period_start).days + 1)]
         all_atts = attr_name.split(";")
[…]
