### 2020-07-15 e6bdb3f
Updated the requirements with versions for PyQt5 and PyQt4.
