### 2020-07-15 benjaminbolling/DynaGUI#3 @benjaminbolling
A requirements.txt file has now been included, let me know if it is sufficient or needs more work.
