+++ requirements.txt
@@ -9,6 +9,7 @@ fnmatch
 numexpr
 math
 functools
+subprocess
 PyQt5==5.13.1 or PyQt4==4.11.4
 pyqtgraph==0.10.0
 matplotlib==3.1.1

[…]
