### 2020-07-15 d387cd1
Updated figures for README.md
