+++ README.md
@@ -24,9 +24,9 @@ The package has then evolved to have the ability to analyse data from any file c
         <td>
             <img alt="DynaGUI Launcher" src="figure2.png">
         </td>
-	DynaGUI Launcher
     </tr>
 </table>
+DynaGUI Launcher
 <table>
     <tr>
         <td>
