+++ README2.md
@@ -1,82 +1,0 @@
-![OpenOpps Platform](assets/images/logo.png)
-
-Badge | Status
---- | ---
-**Circle CI Build** | [![Build Status](https://circleci.com/gh/openopps/openopps-platform.svg?style=shield&circle-token=fc05032b464335e120cd4133f124a3b694bc5d2b)](https://circleci.com/gh/openopps/openopps-platform)
-**Code Climate** | [![Code Climate](https://codeclimate.com/github/openopps/openopps-platform/badges/gpa.svg)](https://codeclimate.com/github/openopps/openopps-platform)
-**Current Version** | [![Current Version](https://img.shields.io/badge/release-v9.7.0-0e5487.svg)](https://github.com/openopps/openopps-platform/releases)
-
----
-
-[Open Opportunities](http://www.digitalgov.gov/join-digitalgov/open-opportunities-in-digitalgov/) is creating a network of federal employees that will lead to a more effective, efficient, responsive government by sharing skills and collaborating on projects that support the mission of an agency or the government as a whole.
-
-OpenOpps is the open source platform that supports the Open Opportunities program.  As open source software, it can be deployed by state or local governments or in the private sector.  It is free to use by anyone who can has technical operations staff and passion to start a new program, and we welcome everyone who would like to contribute to improving the platform.
-
-Here’s how Open Opportunities works: You come up with a great idea that would help your workplace, but you and your coworkers don’t have the skills to execute it. Rather than bringing in a contractor, you can post your project idea to Open Opportunities, where other people within your organization can see it and then work on it, if they’d like.
-
-Video: [Open Opportunities Testimonial from DigitalGov Search Manager](https://www.youtube.com/watch?v=2FelzCR7pLk&feature=youtu.be)
-
-Open Opportunities can benefit all organizations, but it’s especially useful for those with distributed teams and whose employees have diverse skill sets.  Using Open Opportunities, you and your colleagues can work on projects together, get to know each other better, and develop skills you don’t use as often as you’d like.
-
-## Screenshots
-
-<table>
-    <tr>
-        <td>
-            <img alt="open opps 1" src="docs/screenshots/open-opps-01.png">
-        </td>
-        <td>
-            <img alt="open opps 2" src="docs/screenshots/open-opps-02.png">
-        </td>
-        <td>
-            <img alt="open opps 3" src="docs/screenshots/open-opps-03.png">
-        </td>
-    </tr>
-</table>
-
-## How you can help
-
-* Use our [community test server](https://openopps-staging.app.cloud.gov) and write up [issues](https://github.com/openopps/openopps-platform/issues)
-* Join our [Google Group](https://groups.google.com/forum/#!forum/openopps-platform) and introduce yourself, please ask any general questions here -- how the product is supposed to work, usability issues, technical questions, whatever you want to know, we want to hear about.  If you are interested in using Open Opportunities for your organization, let us know.
-* Read our [Contribution Guide][contributing] -- everything you do with us is a contribution to the public domain, there are also some tips for getting started
-* If you are a developer, [install](INSTALL.md) the software for development
-* Dive into our [help wanted issues](https://github.com/openopps/openopps-platform/labels/help%20wanted) -- there are both dev and design opportunities there
-
-## More Project Details
-
-Problem Statement: Our workforce often lacks talent and skills where and when they are needed.
-
-- working across organizational boundaries is difficult
-- professional development opportunities are scarce
-- only 35% of federal employees believe creativity and innovation are rewarded (2014 Federal Employee Viewpoint Survey)
-
-Solution: provide a software platform that enables a program with very small central staff and program leaders across government that provide professional development opportunities
-
[…]
