### 2020-07-15 benjaminbolling/DynaGUI#5 @benjaminbolling
The README file has been updated. @Chilipp Since you opened this issue, I leave it up to you to decide if this issue is to be considered as closed or not.
