### 2020-07-16 2166f39
Added 'Example usages' to the README file.
