### 2020-07-16 benjaminbolling/DynaGUI#7 @benjaminbolling
I have put some example usage in the README file, please let me know if it is sufficient or not.
