### 2020-07-16 ca344bb
Create CONTRIBUTING.md
