+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Welcome!
+Thank you for considering contributing to this open source project. To do a contribution, please follow the following steps:
+
+1. Create your own branch and develop the code and your contribution.
+2. If your contribution is useful for the community, proceed with a merge request to the master (community) branch.
+3. Document your contribution (API documentation), write a short description of what you have done and its potential use-cases.
+4. Start the pull request.
+
+## Issues
+If you experience any issue, please submit it as an issue ticket (but please ensure that it will not result in any duplicate) and label it as a 'bug'.
+
+## Feature request
+If there is a feature that you would like to see in a later version, please submit it as an issue ticket and label it as an 'enhancement'.
+
+## Contact details
+For direct questions or enquiries, you may write an email with the title "DynaGUI" to:
+Email: benjaminbolling@icloud.com
