### 2020-07-16 b976e18
Update CONTRIBUTING.md
