### 2020-07-16 benjaminbolling/DynaGUI#9 @benjaminbolling
I have created the CONTRIBUTING.md file, please check if you consider it as sufficient or not.
