### 2020-07-16 benjaminbolling/DynaGUI#8 @benjaminbolling
This might take a little time, but I am working on it.
