### 2020-07-16 47dc123
Added 'Installation procedure' to the README file.
