### 2020-07-16 0a545a0
Now citing NumPy and MatplotLib in the paper.
