+++ paper.bib
@@ -12,20 +12,20 @@
     Author = {Chaize, J. M. and Gotz, A. and Klotz, W-D. and Meyer, J. and Perez, M. and Taurel, E.},
     title = {{TANGO} - an object oriented control system based on {CORBA}},
     journal = {ICALEPCS'99},
-    year = 1999
+    year = {1999}
 }
 
 @webpage{PyQtReference,
     Author = {{Riverbank Computing Limited}},
     title = {{PyQt5}: {P}ython bindings for the {Q}t cross platform {UI} and application toolkit},
     URL = {https://www.riverbankcomputing.com/software/pyqt/},
-    year = 2016
+    year = {2016}
 }
 
 @webpage{pyepicsReference,
     title = {{PyEpics}: {I}nterface for the {C}hannel {A}ccess ({CA}) library of the {EPICS} {C}ontrol {S}ystem to the {P}ython {P}rogramming language},
     URL = {https://cars9.uchicago.edu/software/python/pyepics3/},
-    year = 2014
+    year = {2014}
 }
 
 @article{pandasReference,
@@ -34,4 +34,27 @@
     journal={Python for High Performance and Scientific Computing},
     volume={14},
     year={2011}
+}
+
+@ARTICLE{numpy,  
+    title={The NumPy Array: A Structure for Efficient Numerical Computation},   
+    author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},  
+    journal={Computing in Science   Engineering},   
+    volume={13},  
+    year={2011},  
+    number={2},  
+    pages={22-30}
+}
+
+@Article{matplotlib,
+    author    = {Hunter, J. D.},
+    title     = {Matplotlib: A 2D graphics environment},
+    journal   = {Computing in Science \& Engineering},
+    volume    = {9},
+    number    = {3},
+    pages     = {90--95},
+    abstract  = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.},
+    publisher = {IEEE COMPUTER SOC},
+    doi       = {10.1109/MCSE.2007.55},
+    year      = {2007}
 }
