### 2020-07-16 66bf004
Added 'Comparison to other packages' to the README file.
