### 2020-07-16 #1942 @benjaminbolling
I have addressed issues 3, 4, 5,7, 8, 9.

Installation instructions: Added to the README file.

State of the field: I am unaware of currently existing Python-based package for constructing dynamic GUI:s (without any programming), and I have added this also to the README file.

Left to do:
- Create some automated tests/steps described so that the functionality of the software can be verified
- Add API documentation

The last part might sadly take a little bit longer time, but I am working on it.
