### 2020-07-17 4418f41
minor bug fixes during test runs and writing of a test procedure. Test procedure written, see 'test.txt'.
