+++ dynagui_files/DynaGUI_NV.py
@@ -66,7 +66,7 @@ class Dialog(QtGui.QWidget):
                                         #'LEBT_Sol-01_Current',
                                         #'LEBT_Sol-02_Current'
                                         ]
-            elif ctrl_library == "Finance":
+            elif self.ctrl_library == "Finance":
                 self.stocknames = ['AAPL',
                                    'TSLA',
                                    'FB',
@@ -200,7 +200,7 @@ class Dialog(QtGui.QWidget):
 
     def savebtnclicked(self):
         # To save the configuration
-        yesorno, nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File')
+        nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File', "", "DynaGUI NV file (*.dg2)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
         else:
@@ -222,7 +222,7 @@ class Dialog(QtGui.QWidget):
 
     def loadbtnclicked(self):
         # File loading initialization
-        nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File')
+        nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File')[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled loading configuration.")
         else:
@@ -620,7 +620,6 @@ class Dialog(QtGui.QWidget):
             if hasattr(val, "__len__"):
                 scalars.append(1)
             else:
-                print("scalar")
                 scalars.append(0)
         if sum(scalars) == 0:
             self.scalarflag = 1 # it is a scalar
@@ -1318,6 +1317,8 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
             self.archivemode = 1
         self.contWidget.loadbtn.clicked.connect(self.loadclick)
         self.contWidget.savebtn.clicked.connect(self.saveclick)
+        self.contWidget.loadbtn.setEnabled(False)
+        self.contWidget.savebtn.setEnabled(False)
         self.contWidget.plotsettingsbtn.clicked.connect(self.PlotSettings)
         self.contWidget.showhidelegends.clicked.connect(self.showhidelegend)
         self.contWidget.resetbtn.clicked.connect(self.reset)
@@ -1761,7 +1762,7 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
             self.contWidget.plotbtn.setText('Start Plotting')
             self.pyqtgraphtimer.stop()
             if str(item) == 'From File':
-                nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File')
+                nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File', "", "Plot files (*.dgplot)")[0]
                 if any(map(lambda x: any(x), nameoffile)):
                     if isinstance(nameoffile, tuple):
                         nameoffile = str(nameoffile[0])
@@ -1995,10 +1996,9 @@ class PyQtGraphPlotter(QtGui.QMainWindow):
     def saveclick(self):
         options = QtGui.QFileDialog.Options()
         options |= QtGui.QFileDialog.DontUseNativeDialog
-        fileName = QtGui.QFileDialog.getSaveFileName(self,"QtGui.QFileDialog.getSaveFileName()","","All Files (*);;Text Files (*.txt)", options=options)
+        fileName = QtGui.QFileDialog.getSaveFileName(self, 'Save to File',"","DynaGUI plot files (*.dgplot)", options=options)[0]
[…]
