### 2020-07-17 #1942 @benjaminbolling
@Chilipp @pythonpanda2 @majensen : I have written a test procedure, let me know if you find it sufficient or not. The API documentation will, as stated before, take a little while (ca. a week). You can start testing it anyways if you'd want to, as it seems to me as if all other issues and comments have now been addressed. Thank you all for the patience and the comments, and have a nice weekend!
