### 2020-07-17 #1942 @majensen
@benjaminbolling thanks for this work. A tip, if you rename your test.txt to test.md, then GitHub will render it in a nice form.

@pythonpanda2 - based on the work done, can you check off the items in your review that are satisfied?

thanks everyone
