### 2020-07-18 #1942 @benjaminbolling
thanks @majensen, I have renamed test.txt to test.md as you suggested.
