### 2020-08-01 benjaminbolling/DynaGUI#4 @pythonpanda2
Thanks a lot for the explanation. I can clearly see some of the explanation in requirements. So i will close the issue and add further comments in the requirements issue.
