### 2020-08-01 #1942 @pythonpanda2
@majensen  @benjaminbolling  I have closed this issue . https://github.com/benjaminbolling/DynaGUI/issues/2
