### 2020-08-01 benjaminbolling/DynaGUI#3 @pythonpanda2
Thanks a lot for uploading the conda environment. It requires a bit more clean up. So i tried to clean it up and try to install the packages in to a local environments. Please see some of the changes below.

I have revamped the requirement.  please find the cleaned up version attached below.
[requirements_modified.txt](https://github.com/benjaminbolling/DynaGUI/files/5011088/requirements_modified.txt)

Do rename it as yml file.   A list of changed  are detailed below:
I went with the PyQt5==5.13.1
Fixed the typo :  PyTango==9.3.2 --> pytango==9.3.2

 Ideally i would like the users to install the package to their own conda enviroments to avoid conflicts with default packages. Hence I added the channels and environment specifications.

Python by default comes with a number of built-in packages which are not required to be specified in the yml file (https://docs.python.org/3/library/) .  I have removed the following :

**fnmatch, os , subprocess , datetime, sys,  random, platform, math and functools**

I ran the installation as follows

`conda env create --file requirements_modified.yml  `

Most of the packages were installed seamlessly under two different settings (Mac and Linux) . Unfortunately, the PyTango failed in both setting. As you are the expert, I would urge take a look at this further and hopefully we can have a working conda enviroment setup.
