### 2020-08-01 #1942 @pythonpanda2
@majensen  @benjaminbolling  It would be good to see if we can resolve this : https://github.com/benjaminbolling/DynaGUI/issues/3
