### 2020-08-03 benjaminbolling/DynaGUI#9 @Chilipp
hi @benjaminbolling. If it's sufficient for you, it's sufficient for me. Thanks! JOSS does not say anything about the quality or necessary contents of this Guide.

One comment, though: People raising issues in your Github repository usually do not have the rights to set _labels_ (e.g. _bug_ or _enhancement_). This is something the code maintainer has to do.
