### 2020-08-03 benjaminbolling/DynaGUI#7 @Chilipp
Sorry @benjaminbolling, but I can't see how this is an example usage, but probably there is a misunderstanding. The Example Usage should guide the user through an example, something like a _Getting started_ section. What you wrote down here is about how _you_ used and developed the package. It should however be an **example** about how the _user_ can use, and get started with your software.
