### 2020-08-08 benjaminbolling/DynaGUI#3 @pythonpanda2
https://github.com/openjournals/joss-reviews/issues/1942#issuecomment-667574261
