### 2020-08-17 #1942 @majensen
@benjaminbolling Have you been able to take care of benjaminbolling/DynaGUI#3 per @pythonpanda2 ?
