### 2020-08-21 benjaminbolling/DynaGUI#3 @benjaminbolling
In order to be able to install PyTango, the user must first install TANGO Controls: https://www.tango-controls.org/downloads/

After this, also Boost.Python must be installed:
pip install boost-py

Question is how to proceed now. There are some lightsources around the world that utilise TANGO Controls, however, it is not required/essential in order for the DynaGUI package to work.
