### 2020-08-21 benjaminbolling/DynaGUI#3 @benjaminbolling
@pythonpanda2 Thank you for the cleaned up version of the requirements, I have pushed it to the master branch.
