### 2020-08-21 34ea5d8
Updated the title of a person in acknowledgements in the paper and the requirements file (thank you pythonpanda2).
