+++ requirements.yml
@@ -1,0 +1,31 @@
+name: dynagui
+channels:
+  - conda-forge
+  - defaults
+  - tango-controls
+  - anaconda
+dependencies:
+# Core dependencies for dynagui
+  - python==3.7.*
+  - numpy
+  - time
+  - numexpr
+  - pyqtgraph==0.10.0
+  - matplotlib==3.1.1
+#For EPICS:
+  - pyepics==3.4.0
+  - requests==2.22.0
+  - jsonschema==3.1.1
+#For Finance:
+  - pandas==0.25.3
+  - pandas-datareader==0.8.1
+  - pip:
+    #For Tango:
+    - pytango==9.3.2
+    - PyQt5==5.13.1
+
+
+
+      
+      
+      
