### 2020-08-21 benjaminbolling/DynaGUI#7 @benjaminbolling
Aha, I see. My mistake, I will update it during the weekend and turn it into a more proper _getting started_. Thank you @Chilipp
