### 2020-08-21 benjaminbolling/DynaGUI#3 @pythonpanda2
@majensen @benjaminbolling  It seems the fix should be straight forward. If the tango has to be installed seperately then would mind adding the instruction for that in the README ?. Then it should be followed by an instruction to create conda enviroment.

You would also need to add **boost-py** with the right version in to the **pip** section of the    requirements.yml . Otherwise it would be incomplete.

Once you are done with this i will be happy to  test the installation.
