### 2020-08-24 c347da5
Separated installation procedure for with or without PyTango usage since the PyTango package requires installation of the TANGO Controls.
