### 2020-08-24 e67f07f
Fixed a formatting issue for the README file
