+++ README.md
@@ -82,9 +82,13 @@ If TANGO is not required by the end user, see steps a.
    b)
 2. Check Python version used with the PIP package manager such that it points to the correct Python version (pip -V).
 3. a) Use PIP to install all packages required, see [requirements](requirements.yml), or use conda to create an environment:
+
     conda env create --file environments.yml
+
    b) Use PIP to install all packages required, see [requirements](requirements_tango.yml), or use conda to create an environment:
+
     conda env create --file environments_tango.yml
+    
 4. If all required Python packages have been successfully installed, the package is ready.
 
 To launch the package, browse to the package's destination in a terminal and execute | python Launcher.py | to launch DynaGUI.
