### 2020-08-24 3ffd1f3
Fixed a formatting issue for the README file
