+++ README.md
@@ -79,7 +79,6 @@ In order to use the PyTango package, the TANGO Controls has to be installed, the
 If TANGO is not required by the end user, see steps a.
 
 1. In order to setup this package, ensure that Python 3.x (3.7 is recommended) is installed on the computer.
-   b)
 2. Check Python version used with the PIP package manager such that it points to the correct Python version (pip -V).
 3. a) Use PIP to install all packages required, see [requirements](requirements.yml), or use conda to create an environment:
 
@@ -88,7 +87,7 @@ If TANGO is not required by the end user, see steps a.
    b) Use PIP to install all packages required, see [requirements](requirements_tango.yml), or use conda to create an environment:
 
     conda env create --file environments_tango.yml
-    
+
 4. If all required Python packages have been successfully installed, the package is ready.
 
 To launch the package, browse to the package's destination in a terminal and execute | python Launcher.py | to launch DynaGUI.
