### 2020-08-24 d1f4ecd
Fixed a formatting issue for the README file
