+++ README.md
@@ -80,11 +80,11 @@ If TANGO is not required by the end user, see steps a.
 
 1. In order to setup this package, ensure that Python 3.x (3.7 is recommended) is installed on the computer.
 2. Check Python version used with the PIP package manager such that it points to the correct Python version (pip -V).
-3. a) Use PIP to install all packages required, see [requirements](requirements.yml), or use conda to create an environment:
+3. a) Use PIP to install all packages required, see [requirements](requirements.yml), or use conda to create the environment:
 
     conda env create --file environments.yml
 
-   b) Use PIP to install all packages required, see [requirements](requirements_tango.yml), or use conda to create an environment:
+   b) Use PIP to install all packages required, see [requirements with PyTango](requirements_tango.yml), or use conda to create the environment:
 
     conda env create --file environments_tango.yml
 
