### 2020-08-24 benjaminbolling/DynaGUI#3 @benjaminbolling
Hi, thank you @pythonpanda2, please have another look when you have time. I separated the requirements and environment into two separate files, one in which PyTango is included and the other without it.
