### 2020-09-08 benjaminbolling/DynaGUI#3 @pythonpanda2
@benjaminbolling @majensen  Just got back from the long weekend. I will try to finish off the tests this week.
