### 2020-09-15 benjaminbolling/DynaGUI#3 @pythonpanda2
@majensen @benjaminbolling  Thanks a lot for the changes. I went back and tried to follow the instructions in the installation section of the README.

When i follow the step 3) i get the follow error

```
UnavailableInvalidChannel: The channel is not accessible or is invalid.
  channel name: epics-controls
  channel url: https://conda.anaconda.org/epics-controls
  error code: 404

You will need to adjust your conda configuration to proceed.
Use `conda config --show channels` to view your configuration's current state,
and use `conda config --show-sources` to view config file locations.
```

Looks like **epics-controls** channel is not listed in anaconda as a channel. I will be available most of this week for quick follow ups.
