### 2020-10-04 a5fc022
Removed Tango and Epics controls from channels in environment files
