+++ environment_tango.yml
@@ -5,6 +5,7 @@ channels:
 dependencies:
   - python
   - boost-py
+  - epics-base
   - functools
   - jsonschema
   - matplotlib
