+++ environment.yml
@@ -4,8 +4,6 @@ channels:
   - nodefaults
 dependencies:
   - python
-  - boost-py
-  - epics-base
   - jsonschema
   - matplotlib
   - numexpr
