### 2020-10-04 benjaminbolling/DynaGUI#3 @benjaminbolling
Dear @pythonpanda2, I haven't received (or seen) any notification about your last comment. I am truly sorry for that, and I have updated the environment files such that epics-controls are no longer part of the channels as it should not need to be a channel (in the Conda environment).

This is strange though, as when I did the last comment on Aug 24th, the environment was set up and seemed to work fine.
