### 2020-10-04 benjaminbolling/DynaGUI#7 @benjaminbolling
@Chilipp Sorry for not responding to this more urgently, I had a lot of things to attend to which kept on pushing this task. Again, sorry for this. I have updated the Readme file such that there is a section User Guide which points the reader towards the UserGuide.pdf file. Is this sufficient or would you prefer to have (parts of) that text in the Readme file directly?
