### 2020-10-05 #1942 @kthyng
@Chilipp and @pythonpanda2 Hi! I am the rotating associate editor in chief here this week trying to help ping stale issues. @Chilipp I see you said just above that you are satisfied with the submission, but you have a lot of unchecked boxes above. Could you either check the boxes or state what are the next issues?
@pythonpanda2 How is your review coming along?

Thanks all!
