### 2020-10-05 #1942 @Chilipp
Hello @kthyng! There seems to be a misunderstanding. No, the items are not yet resolved. My last comment just states that the open points I mentioned are still open. @benjaminbolling just made some changes yesterday and I'll review them as soon as possible
