### 2020-10-06 benjaminbolling/DynaGUI#3 @pythonpanda2
@majensen @benjaminbolling  No problem.

Thanks a lot for making the changes. I went through the installation again.  `Step 3a) conda env create --file environments.yml ` worked fine.

A small typo in  Step 3b) documentation. `conda env create --file environments_tango.yml`  **should be** ` conda env create --file environment_tango.yml`

A quick question. Does this mean that the user should install Tango package on their own before  proceeding with step 3b?
