### 2020-10-07 #1942 @majensen
@kthyng thanks for checking in. Things have started to proceed over at https://github.com/benjaminbolling/DynaGUI/issues/3#issuecomment-704386598
