### 2020-10-15 e9ae2c1
minor updates to some namings
