+++ dynagui_files/DynaGUI_TF.py
@@ -39,28 +39,15 @@ class Dialog(QtGui.QDialog):
                 loadflag = 0
         if loadflag == 0:
             # All device attributes needed (basically which can be TRUE or False)
-            self.listofbpmattributes = ['AGCEnabled',
-                                   'ADCEnabled',
-                                   'EnableADC',
-                                   'InterlockEnabled',
-                                   'Interlock_Enabled',
-                                   'Attribute1',
+            self.listofattributes = ['Attribute1',
                                    'Attribute2',
-                                   'Attribute3',
-                                   'StatusOpen',
-                                   'StatusClosed',
-                                   'OutputOn']
+                                   'Attribute3']
 
-            # Some list of devices. List 3 consists of some random stuff
-            self.devlist = ['r1-101/dia/bpm-02',
-                     'i-s01b/dia/bpl-01',
-                     'i-s04b/dia/bpl-01',
-                     'r3-301l/dia/bpm-01',
-                     'i-s19b/dia/bpl-01',
-                     'r1-109/dia/bpm-03',
-                     'i-s01a/vac/vgmb-01',
-                     'i-s03a/vac/vgmb-01',
-                     'r3-a110711cab08/mag/pspi-02']
+            # Some list of devices
+            self.devlist = ['section1/discipline1/device1',
+                            'section1/discipline1/device2',
+                            'section3/discipline1/device1',
+                            'section3/discipline2/device1']
             self.Nrows = 20
         self.reloadflag = 0
         self.showallhideflag = False
@@ -75,10 +62,10 @@ class Dialog(QtGui.QDialog):
         self.toplayout.addStretch()
 
         # Construct the combobox for the list of attributes
-        self.listofbpmattributeslistbox = QtGui.QComboBox(self)
-        self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
-        self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
-        self.toplayout.addWidget(self.listofbpmattributeslistbox)
+        self.listofattributeslistbox = QtGui.QComboBox(self)
+        self.listofattributeslistbox.addItems(self.listofattributes)
+        self.listofattributeslistbox.currentIndexChanged.connect(self.statuscheck)
+        self.toplayout.addWidget(self.listofattributeslistbox)
 
         # Here we add a button that sets the selected attribute as TRUE for all the BPM:s in the selected ring
         self.enableallbutton = QtGui.QPushButton("Enable all")
@@ -136,7 +123,7 @@ class Dialog(QtGui.QDialog):
             self.bottomlabel.setText("Cancelled save configuration.")
         else:
             file = open(nameoffile, 'w')
-            self.toSave = str('IamaDynaGUIfile' + '\n' + "##IamYourSeparator##\n" + '\n'.join(self.devlist) + '\n' + "##IamYourSeparator##\n" + '\n'.join(self.listofbpmattributes) + '\n' + "##IamYourSeparator##\n" + str(self.Nrows))
+            self.toSave = str('IamaDynaGUIfile' + '\n' + "##IamYourSeparator##\n" + '\n'.join(self.devlist) + '\n' + "##IamYourSeparator##\n" + '\n'.join(self.listofattributes) + '\n' + "##IamYourSeparator##\n" + str(self.Nrows))
             file.write(self.toSave)
             file.close()
             self.bottomlabel.setText("Configuration saved to file.")
@@ -163,17 +150,17 @@ class Dialog(QtGui.QDialog):
                     devlist = splitToLoad[1].split("\n")
                     while("" in devlist): # Get rid of empty strings
                         devlist.remove("")
-                    listofbpmattributes = splitToLoad[2].split("\n")
-                    while("" in listofbpmattributes): # Get rid of empty strings
-                        listofbpmattributes.remove("")
[…]
