### 2020-10-19 benjaminbolling/DynaGUI#7 @Chilipp
Hi @benjaminbolling, apologies for the delay! I am quite busy at the moment, too :sweat_smile:. So what I understand of an _Example usage_ is a section like this: https://psyplot.readthedocs.io/projects/psy-view/en/latest/getting-started.html. It's a guide with some initial data or so that the user can run through to get an initial impression of the software after having it installed. This is also, what the JOSS requirements say:

> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I could not find something comparable in your user guide.

Do you have something like this? Or do you think this is impossible to generate for DynaGUI? Or do you think it's not valuable for DynaGUI (in which case we need to ask the editor, how to proceed)?
