### 2020-10-29 benjaminbolling/DynaGUI#3 @benjaminbolling
I am sorry, I thought I posted a reply last week but apparently I didn't.

I fixed the typo, thanks for noticing it.
Correct, the Tango package has to be installed by the user before proceeding with step 3b.
