### 2020-10-29 #1942 @benjaminbolling
@majensen Sorry for the delay in my reply, I have a lot of work nowadays during the pandemic.

I have replied to [Issue #3](https://github.com/benjaminbolling/DynaGUI/issues/3).

I will deal with [Issue #7](https://github.com/benjaminbolling/DynaGUI/issues/7) this evening.

I am close to finished with the API documentation for [Issue #8](https://github.com/benjaminbolling/DynaGUI/issues/8), and will post here when it is finished.

Again, thanks for all patience and all your efforts put into this!
