### 2020-10-29 ac995cd
Found a bug identifying the random and finance packages in the launcher.
