### 2020-10-29 55279ba
Began writing on the Getting Started section
