### 2020-10-29 benjaminbolling/DynaGUI#7 @benjaminbolling
@Chilipp : I have begun writing a Getting Started section and will soon continue (in a few hours). If you see this comment and you have time to do a quick check, let me know if that is what you meant. 🙂
