### 2020-10-29 32e2a28
Began writing on the Getting Started section
