### 2020-10-29 28bbccf
Minor fix to the name generation of the random attributes of NV
