+++ README.md
@@ -45,10 +45,10 @@ The package can be launched by executing `python Launcher.py` in a terminal from
         </td>
     </tr>
 </table>
-Depending on which packages are installed, different buttons will be enabled and disabled. The package that is included with Python is Random and is hence expected to always load properly, and which we will hence use in this section.
+Depending on which packages are installed, different buttons will be enabled and disabled. The package that is included with Python is Random and is hence expected to always load properly, and which is hence used in the examples in this section. The ideal procedure would be if this tutorial is completed using the Random package and then repeated using the package DynaGUI is intended to be used with (EPICS, Tango or Finance).
 
 ### Part 2: The Boolean Controller (also referred to as TF)
-Clicking on Boolean Controller beneath Random in the launcher will open a widget that looks like this:
+Clicking on Boolean Controller beneath the selected package in the launcher will open a widget that looks like this:
 <table>
     <tr>
         <td>
@@ -66,7 +66,7 @@ Each button represents an individual device. The status of each device with the 
 </table>
 
 ### Part 3: Value Alarms
-Clicking on Value Alarms beneath Random in the launcher will open a widget that looks like this:
+Clicking on Value Alarms beneath the selected package in the launcher will open a widget that looks like this:
 <table>
     <tr>
         <td>
@@ -84,7 +84,7 @@ Each checkbox represents a device together with an attribute, that is, the full 
 </table>
 
 ### Part 4: Plotting (also referred to as NV)
-Clicking on Plotting beneath Random in the launcher will open a widget that looks like the top-left widget in the figure below:
+Clicking on Plotting beneath the selected package in the launcher will open a widget that looks like the top-left widget in the figure below:
 <table>
     <tr>
         <td>
