### 2020-10-29 benjaminbolling/DynaGUI#7 @benjaminbolling
Dear @Chilipp, I have now completed a getting started section. Please let me know if this is what you had in mind.
