### 2020-10-29 #1942 @benjaminbolling
I have now dealt with [Issue #7](https://github.com/benjaminbolling/DynaGUI/issues/7), let's see if this will do. 🙂

Continuing with [Issue #8](https://github.com/benjaminbolling/DynaGUI/issues/8) now.
