+++ dynagui_files/DynaGUI_TF.py
@@ -16,6 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
+# Import all needed Python packages
 try:
     import PyQt5.QtWidgets as QtGui
     from PyQt5 import QtCore
@@ -24,7 +25,9 @@ except:
 import time, sys, platform
 
 class Dialog(QtGui.QDialog):
+    # This is the class in which the DynaGUI window is constructed
     def __init__(self, inp, ctrl_library):
+        # Intialization of start-up parameters and Qt
         QtGui.QDialog.__init__(self)
         self.setWindowTitle("DynaGUI TF")
         self.ctrl_library = ctrl_library
@@ -32,35 +35,37 @@ class Dialog(QtGui.QDialog):
             loadflag = 0
         else:
             try:
-                self.loadfile(inp,1)
+                self.loadfile(inp,1) # 1 means that a file is being loaded before anything was constructed (load on start)
                 self.Nrows
                 loadflag = 1
             except:
                 loadflag = 0
         if loadflag == 0:
-            # All device attributes needed (basically which can be TRUE or False)
+            # All device attributes (which can be 0/1) needed, used for initialization during start-up
             self.listofattributes = ['Attribute1',
                                    'Attribute2',
                                    'Attribute3']
-
-            # Some list of devices
+            # Some list of devices, used for initialization during start-up
             self.devlist = ['section1/discipline1/device1',
                             'section1/discipline1/device2',
                             'section3/discipline1/device1',
                             'section3/discipline2/device1']
             self.Nrows = 20
-        self.reloadflag = 0
+        # Show or hide enable/disable all btn
         self.showallhideflag = False
-
         if ctrl_library == "Randomizer":
             self.devstat = []
             for m in range(len(self.devlist)):
                 self.devstat.append(1)
-
+        # Create the layout of the GUI
+        self.createLayout()
+        # Run the script for generating the dynamical buttons
+        self.getallDevs()
+    def createLayout(self):
+        # Create the layout of the widget.
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
         self.toplayout.addStretch()
-
         # Construct the combobox for the list of attributes
         self.listofattributeslistbox = QtGui.QComboBox(self)
         self.listofattributeslistbox.addItems(self.listofattributes)
@@ -113,11 +118,8 @@ class Dialog(QtGui.QDialog):
         self.updatebutton.clicked.connect(self.statuscheck)
         self.toplayout.addWidget(self.updatebutton)
         self.toplayout.addWidget(self.enableallbutton)
-
-        # Run the script for generating the dynamical buttons
-        self.getallDevs()
-
     def savebtnclicked(self):
+        # Save the current configuration to a file
         nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File', "", "DynaGUI TF files (*.dg1)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
@@ -128,63 +130,60 @@ class Dialog(QtGui.QDialog):
             file.close()
             self.bottomlabel.setText("Configuration saved to file.")
             self.bottomlabel.setToolTip("Saved configuation to file: "+nameoffile)
-
     def loadbtnclicked(self):
+        # The button was clicked to load a file
         nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File', "", "DynaGUI TF files (*.dg1)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled loading configuration.")
         else:
-            self.loadfile(nameoffile,0)
-
+            self.loadfile(nameoffile,0) # 0 means that there is data to be replaced
[…]
