### 2020-10-31 ae1abd5
Updated all DynaGUI scripts a bit, added API documentation to DynaGUI alarms
