+++ dynagui_files/DynaGUI_Alarms.py
@@ -28,10 +28,13 @@ from functools import partial
 from time import sleep
 import sys
 class Dialog(QtGui.QDialog):
+    # This is the class in which the DynaGUI window is constructed
     def __init__(self, inp, ctrl_library):
+        # Intialization of start-up parameters and Qt
         QtGui.QDialog.__init__(self)
         self.ctrl_library = ctrl_library
         self.setWindowTitle("DynaGUI Alarms")
+        # Input 0 means load with standard pre-defined configurations (no file to load to during start-up)
         if inp == 0:
             loadflag = 0
         else:
@@ -42,7 +45,7 @@ class Dialog(QtGui.QDialog):
             except:
                 loadflag = 0
         if loadflag == 0:
-            # List of devices' server domains.
+            # List of some sample channels, some samples to load here if no start-up file was defined above
             if self.ctrl_library == "Tango":
                 self.devdoms = ['section1/discipline1/device1/temp',
                                 'section1/discipline1/device2/press',
@@ -104,10 +107,17 @@ class Dialog(QtGui.QDialog):
                                 100,
                                 1000]
             self.Nrows = 20
+        # Reloadflag is used for stating if there are pre-existing control buttons; 1 == rmv them, 0 == nothing to remove which is only during start-up
         self.reloadflag = 0
-        self.maxsize = 0
         self.timerinterval = 30 # seconds
-
+        # Create the layout
+        self.createLayout()
+        # Run the script for generating the dynamical buttons with the channels defined
+        self.getallDevs()
+        self.timer = QtCore.QTimer()
+        self.timer.timeout.connect(self.statuscheck)
+    def createLayout(self):
+        # Create the layout
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
         self.toplayout.addStretch()
@@ -172,21 +182,16 @@ class Dialog(QtGui.QDialog):
         self.startstopbtn.setStyleSheet('QPushButton {background-color: maroon; color: white}')
         self.startstopbtn.clicked.connect(self.startstopclicked)
         self.toplayout.addWidget(self.startstopbtn)
-        # Run the script for generating the dynamical buttons
-        self.getallDevs()
-
-        self.timer = QtCore.QTimer()
-        self.timer.timeout.connect(self.statuscheck)
-
     def selectallbtnclicked(self):
+        # Set all alarms to be checked
         for item in self.groupBox.findChildren(QtGui.QCheckBox):
             item.setChecked(True)
-
     def unselectallbtnclicked(self):
+        # Set all alarms to be unchecked
         for item in self.groupBox.findChildren(QtGui.QCheckBox):
             item.setChecked(False)
-
     def savebtnclicked(self):
+        # Save the DynaGUI configuration
         nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File', "", "DynaGUI Alarms file (*.dg3)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
@@ -197,56 +202,55 @@ class Dialog(QtGui.QDialog):
             file.close()
             self.bottomlabel.setText("Configuration saved to file.")
             self.bottomlabel.setToolTip("Saved configuation to file: "+nameoffile)
-
     def loadbtnclicked(self):
+        # The button was clicked to load a file
         nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File', "", "DynaGUI Alarms file (*.dg3)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled loading configuration.")
         else:
             self.loadfile(nameoffile,0)
-
     def loadfile(self,nameoffile,inp2):
[…]
