### 2020-10-31 073961a
Updated DynaGUI alarms API documentation. Added API documentation to the Launcher.
