+++ Launcher.py
@@ -16,7 +16,9 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
+# We create here a simple list which contains if an import of a package failed. 0 = no fail, 1 = fail
 packagefailure = [0,0,0,0,0] # [ basic, tango, epics, random, finance]
+# Try importing required packages. If any of these fails, DynaGUI cannot be initialized
 try:
     import os, platform, sys, time, datetime, fnmatch, numexpr, math, json
     import numpy as np
@@ -35,80 +37,117 @@ except:
     except:
         packagefailure[0] = 1
 if packagefailure[0] == 1:
-    print("Failed to import basic packages.")
+    # At least one required package import failed, don't try importing the others and send this message to the user in the terminal
+    print("Failed to import basic packages, DynaGUI cannot launch.")
     packagefailure = [1,1,1,1,1] # [ basic, tango, epics, random, finance]
 elif packagefailure[0] == 0:
+    # The requireded packages succeeded importing, try with the ones for the different DynaGUI control libraries
     try: # Tango
         import PyTango as PT
     except:
-        print("Failed to import PyTango")
+        print("Failed to import PyTango, Tango control library in DynaGUI is hence disabled")
         packagefailure[1] = 1
     try: # EPICS
         import epics
     except:
-        print("Failed to import Epics")
+        print("Failed to import Epics, EPICS control library in DynaGUI is hence disabled")
         packagefailure[2] = 1
     try: # EPICS
         import requests
     except:
-        print("Failed to import Requests")
+        print("Failed to import Requests, EPICS control library in DynaGUI is hence disabled")
         packagefailure[2] = 1
     try: # Randomizer
         import random
     except:
-        print("Failed to import Random")
+        print("Failed to import Random, Random control library in DynaGUI is hence disabled")
         packagefailure[3] = 1
     try: # Finance
         from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
         from matplotlib.backends.backend_qt4agg import NavigationToolbar2QT as NavigationToolbar
     except:
-        print("Failed to matplotlib backends package")
+        print("Failed to matplotlib backends package, Finance control library in DynaGUI is hence disabled")
         packagefailure[4] = 1
     try: # Finance
         from pandas.plotting import register_matplotlib_converters
         import pandas_datareader as pdr
     except: # Finance
-        print("Failed to import Pandas Finance package")
+        print("Failed to import Pandas Finance package, Finance control library in DynaGUI is hence disabled")
         packagefailure[4] = 1
 
 class Dialog(QtGui.QWidget):
+    # This is the class in which the DynaGUI window is constructed
     def __init__(self, packagefailure,parent=None):
         super(Dialog, self).__init__(parent)
         QtGui.QDialog.__init__(self)
+        # Create the layout
+        self.createLayout()
+        # Disable control libraries that were unsuccessful during loading
+        if packagefailure[0] == 1:
+            self.viewdataNV.setEnabled(False)
+            self.viewdataNV.setToolTip("Import of required general Python packages failed.")
+        else:
+            self.viewdataNV.setEnabled(False)
[…]
