### 2020-11-08 benjaminbolling/DynaGUI#3 @pythonpanda2
Thanks a lot. I will have a look at the issue this week.
