### 2020-11-11 benjaminbolling/DynaGUI#3 @pythonpanda2
Thanks a lot for making the changes. I will close the issue.
