### 2020-11-16 #1942 @benjaminbolling
Good morning all, and thanks again. Can you confirm what is left to be done? I think I am ready to push the changes in terms of the last part of the API documentation, [Issue #8](https://github.com/benjaminbolling/DynaGUI/issues/8).

Let me know what else is left to do. Thanks again all, and sorry for taking so much time.
