### 2020-11-18 #1942 @pythonpanda2
> @benjaminbolling @Chilipp @pythonpanda2 - as we're coming up on the one year anniversary of the submission, I thought I would see if we're still on track. I think @benjaminbolling is making some progress. @Chilipp I think you are still with us. @pythonpanda2 can you ping back on this thread as to whether you are still following this review and able to pick it up again?
> thanks all

I think closed the last issued i had opened. But i will go through the checklist once more by early next week.
