### 2020-11-22 benjaminbolling/DynaGUI#8 @Chilipp
Hey @benjaminbolling! thanks for adding some comments to the code. The question is a bit here, what JOSS exactly requires as an API documentation. A common format of an API reference would be something like https://psyplot.readthedocs.io/projects/psy-view/en/latest/api/psy_view.html or http://xarray.pydata.org/en/stable/api.html. From my perspective, I think your approach is well enough as your main focus is on the usage of the GUI, and not on the Python API. But I'll ask the editor about this
