### 2020-11-22 #1942 @Chilipp
Thanks for your patience and apologies again for the delay @majensen and @benjaminbolling!

I went through all open points from the checklist for JOSS and understand them as fulfilled. So I recommend this paper to be accepted.
