### 2020-11-22 9b13b78
Added API documentation to DynaGUI NV
