+++ dynagui_files/DynaGUI_NV.py
@@ -25,11 +25,15 @@ except:
     from PyQt4 import QtCore, QtGui
 
 class Dialog(QtGui.QWidget):
+    # The QWidget Class GUI for the main view
     def __init__(self, ctrl_library, inp, parent=None):
+        # Initialization of the GUI with a control library and an input
+        # Inp = 0 --> No file to load, Inp = string --> file to load
         super(Dialog, self).__init__(parent)
         QtGui.QDialog.__init__(self)
         self.setWindowTitle("DynaGUI NV - "+str(ctrl_library))
         self.ctrl_library = ctrl_library
+        # Here we check if a file is to be loaded or not during start-up
         if inp == 0:
             loadflag = 0
         else:
@@ -39,8 +43,10 @@ class Dialog(QtGui.QWidget):
                 loadflag = 1
             except:
                 loadflag = 0
+        # No files have been defined so we start with a pre-defined configuration
         if loadflag == 0:
-            if self.ctrl_library == "Tango": # Some MAX IV Laboratory devices and attributes
+            # Check which control library we use
+            if self.ctrl_library == "Tango":
                 self.devlist = ['section1/discipline1/device1',
                                 'section1/discipline1/device2',
                                 'section3/discipline1/device1',
@@ -81,11 +87,10 @@ class Dialog(QtGui.QWidget):
                                 "by",
                                 "Rows"]
             self.Nrows = 15
-        self.reloadflag = 0
+        self.reloadflag = 0 # Reloadflag is a global flag to check if we reload the entire GUI or just update/get the value(s)
         self.maxsize = 0
-        self.toSpecminutes = 15
-        self.toSpecupdateFrequency = 2
-        self.showallhideflag = False
+        self.toSpecminutes = 15 # Global setting for how many minutes to span plotting elements over
+        self.toSpecupdateFrequency = 2  # Global setting for the plotting frequence of plotting elements
 
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
@@ -93,17 +98,17 @@ class Dialog(QtGui.QWidget):
 
         # Construct the combobox for the list of attributes
         if self.ctrl_library == "Tango":
-            self.listofbpmattributeslistbox = QtGui.QComboBox(self)
-            self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
-            self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
-            self.toplayout.addWidget(self.listofbpmattributeslistbox)
+            self.listofElementAttributesListBox = QtGui.QComboBox(self)
+            self.listofElementAttributesListBox.addItems(self.listofbpmattributes)
+            self.listofElementAttributesListBox.currentIndexChanged.connect(self.statuscheck)
+            self.toplayout.addWidget(self.listofElementAttributesListBox)
         elif self.ctrl_library == "Randomizer":
-            self.listofbpmattributeslistbox = QtGui.QComboBox(self)
-            self.listofbpmattributeslistbox.addItems(self.listofbpmattributes)
-            self.listofbpmattributeslistbox.currentIndexChanged.connect(self.statuscheck)
-            self.toplayout.addWidget(self.listofbpmattributeslistbox)
-        # elif self.ctrl_library == "EPICS": DO NOT add, EPICS uses PV:s and not devices with attributes
-        # elif self.ctrl_library == "Finance": DO NOT add, makes no sense.
+            self.listofElementAttributesListBox = QtGui.QComboBox(self)
+            self.listofElementAttributesListBox.addItems(self.listofbpmattributes)
+            self.listofElementAttributesListBox.currentIndexChanged.connect(self.statuscheck)
[…]
