### 2020-11-22 benjaminbolling/DynaGUI#8 @benjaminbolling
Dear @Chilipp, thank you so much for your comments. I have elaborated the API documentation for the DynaGUI NV as well now, and await @majensen (assuming he is the editor) for comments on this. Hopefully this tagging works! :)
