### 2020-11-22 #1942 @pythonpanda2
I have reviewed the checklist. I would  recommend the DynaGUI research software article  for publication in JOSS.
