### 2020-11-22 benjaminbolling/DynaGUI#8 @majensen
Hi @benjaminbolling: I think this is good. What I would do is the following one simple trick: for all the methods where you have a comment following the `def`, such as
```
def __init__(self, ctrl_library, inp, parent=None):
      # Initialization of the GUI with a control library and an input
```

triple-quote the comment instead:

```
def __init__(self, ctrl_library, inp, parent=None):
     """Initialization of the GUI with a control library and an input"""
```
This will turn that comment into the doc line for the method. Then it will be visible to things like `pydoc`, and can also be parsed for fancier rendering as in the pages that @Chilipp referred to.
If we do that, I think we can call it good.

(note triple-quotes leave the line breaks in, so for example, do
```
def myfunc(self):
    """This is a multi-line docstring:
Second line
Third line
""""
```
and this will preserve the line-breaks in the downstream formatters)
