### 2020-11-23 c42d7df
Updated API documentation: All comments following and hence describing a def use triple-quote (""") instead  of a hashtag, turning them into proper doc line for the method.
