+++ dynagui_files/DynaGUI_TF.py
@@ -16,18 +16,18 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-# Import all needed Python packages
+# Import PyQt5 or PyQt4
 try:
     import PyQt5.QtWidgets as QtGui
     from PyQt5 import QtCore
 except:
     from PyQt4 import QtCore, QtGui
-import time, sys, platform
+
 
 class Dialog(QtGui.QDialog):
-    # This is the class in which the DynaGUI window is constructed
+    """This is the class in which the DynaGUI window is constructed."""
     def __init__(self, inp, ctrl_library):
-        # Intialization of start-up parameters and Qt
+        """Intialization of start-up parameters and Qt."""
         QtGui.QDialog.__init__(self)
         self.setWindowTitle("DynaGUI TF")
         self.ctrl_library = ctrl_library
@@ -63,7 +63,7 @@ class Dialog(QtGui.QDialog):
         # Run the script for generating the dynamical buttons
         self.getallDevs()
     def createLayout(self):
-        # Create the layout of the widget.
+        """Create the layout of the widget."""
         # Construct the toplayout and make it stretchable
         self.toplayout = QtGui.QVBoxLayout(self)
         self.toplayout.addStretch()
@@ -120,7 +120,7 @@ class Dialog(QtGui.QDialog):
         self.toplayout.addWidget(self.updatebutton)
         self.toplayout.addWidget(self.enableallbutton)
     def savebtnclicked(self):
-        # Save the current configuration to a file
+        """Save the current configuration to a file."""
         nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File', "", "DynaGUI TF files (*.dg1)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
@@ -132,14 +132,14 @@ class Dialog(QtGui.QDialog):
             self.bottomlabel.setText("Configuration saved to file.")
             self.bottomlabel.setToolTip("Saved configuation to file: "+nameoffile)
     def loadbtnclicked(self):
-        # The button was clicked to load a file
+        """The button was clicked to load a file."""
         nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File', "", "DynaGUI TF files (*.dg1)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled loading configuration.")
         else:
             self.loadfile(nameoffile,0) # 0 means that there is data to be replaced
     def loadfile(self,nameoffile,inp2):
-        # Load a previously saved configuration
+        """Load a previously saved configuration."""
         file = open(nameoffile, 'r')
         splitToLoad = file.read()
         splitToLoad = splitToLoad.split("##IamYourSeparator##")
@@ -177,7 +177,7 @@ class Dialog(QtGui.QDialog):
             if inp2 == 0:
                 self.bottomlabel.setText("Not a DynaGUI file - missing identifier.")
     def enableallbuttonclicked(self):
-        # Set the attribute selected to true for all channels
+        """Set the attribute selected to true for all channels."""
         if ctrl_library == "Randomizer":
             self.devstat.clear()
         for item in self.buttonGroup.buttons():
@@ -197,7 +197,7 @@ class Dialog(QtGui.QDialog):
                 self.devstat.append(1)
         self.statuscheck()
     def killdynamicbuttongroup(self):
-        # Destroy / kill all buttons currently constructed in the buttongroup.
+        """Destroy / kill all buttons currently constructed in the buttongroup."""
         self.bottomlabel.setText(str("Loading " + str(self.listofattributeslistbox.currentText()) + " statuses..."))
         for i in reversed(range(self.sublayout.count())):
             item = self.sublayout.itemAt(i)
[…]
