+++ dynagui_files/DynaGUI_NV.py
@@ -16,7 +16,8 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-# Import all needed Python packages
+
+# Import PyQt5 or PyQt4
 try:
     import PyQt5.QtWidgets as QtGui
     import PyQt5.QtGui as QtGui2
@@ -25,10 +26,10 @@ except:
     from PyQt4 import QtCore, QtGui
 
 class Dialog(QtGui.QWidget):
-    # The QWidget Class GUI for the main view
+    """The QWidget Class GUI for the main view """
     def __init__(self, ctrl_library, inp, parent=None):
-        # Initialization of the GUI with a control library and an input
-        # Inp = 0 --> No file to load, Inp = string --> file to load
+        """Initialization of the GUI with a control library and an input:
+        Inp = 0 --> No file to load, Inp = string --> file to load """
         super(Dialog, self).__init__(parent)
         QtGui.QDialog.__init__(self)
         self.setWindowTitle("DynaGUI NV - "+str(ctrl_library))
@@ -194,7 +195,7 @@ class Dialog(QtGui.QWidget):
         self.getallDevs()
 
     def savebtnclicked(self):
-        # To save the configuration
+        """Btn clicked to save the configuration """
         nameoffile = QtGui.QFileDialog.getSaveFileName(self, 'Save to File', "", "DynaGUI NV file (*.dg2)")[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled save configuration.")
@@ -216,7 +217,7 @@ class Dialog(QtGui.QWidget):
             self.bottomlabel.setToolTip("Saved configuation to file: "+nameoffile)
 
     def loadbtnclicked(self):
-        # Select what file to load and run the loadfile command with the name of the file
+        """Select what file to load and run the loadfile command with the name of the file. """
         nameoffile = QtGui.QFileDialog.getOpenFileName(self, 'Load File')[0]
         if not nameoffile:
             self.bottomlabel.setText("Cancelled loading configuration.")
@@ -224,10 +225,7 @@ class Dialog(QtGui.QWidget):
             self.loadfile(nameoffile,0)
 
     def loadfile(self,nameoffile,inp2):
-        # The actual loading of the file, inputs are name of file and a boolean
-        # stating if the function is called during start-up of the GUI (=1) or
-        # if it has been called by the load file button (0), in which case there
-        # are dynamic objects in the GUI which have to be removed (incl. refs)
+        """The actual loading of the file, inputs are name of file and a boolean stating if the function is called during start-up of the GUI (=1) or if it has been called by the load file button (0), in which case there are dynamic objects in the GUI which have to be removed (incl. refs). """
         file = open(nameoffile, 'r')
         splitToLoad = file.read()
         splitToLoad = splitToLoad.split("##IamYourSeparator##")
@@ -278,7 +276,7 @@ class Dialog(QtGui.QWidget):
                 self.bottomlabel.setText("Not a DynaGUI file - missing identifier.")
 
     def killdynamicbuttongroup(self):
-        # Destroy all objects currently constructed in the buttongroup.
+        """Destroy all objects currently constructed in the buttongroup."""
         if self.ctrl_library == "EPICS":
             self.bottomlabel.setText(str("Loading defined PV statuses..."))
         elif self.ctrl_library == "Finance":
@@ -297,8 +295,7 @@ class Dialog(QtGui.QWidget):
             button.deleteLater()
 
     def getallDevs(self):
-        # This function is made to get all the devices defined and construct the
-        # dynamic buttongroup with connections to all device.
+        """This function is made to get all the devices defined and construct the dynamic buttongroup with connections to all device."""
 
[…]

[…]
