### 2020-11-23 69f9c1f
Updated API documentation: All comments following and hence describing a def use triple-quote (""") instead  of a hashtag, turning them into proper doc line for the method for the DynaGUI Launcher.
