### 2020-11-23 benjaminbolling/DynaGUI#8 @benjaminbolling
Good morning! Thank you for your comments/suggestions, I have implemented these changes to all the Python files.
