### 2020-11-24 #1942 @majensen
Thanks everyone. I surely appreciate all the work and tenacity on this review. I will do some editorial checks on the paper, which may result in minor pull request. Once that is complete, @benjaminbolling, I'll ask you to create a public archive of the repo using Zenodo or similar. After that we can summon the associate editors-in-chief by making our recommendation.
