### 2020-12-13 82c0b5f
Added DOI reference to to NumPy in paper.bib.
