+++ paper.bib
@@ -36,14 +36,15 @@
     year={2011}
 }
 
-@ARTICLE{numpy,  
-    title={The NumPy Array: A Structure for Efficient Numerical Computation},   
-    author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},  
-    journal={Computing in Science   Engineering},   
-    volume={13},  
-    year={2011},  
-    number={2},  
-    pages={22-30}
+@ARTICLE{numpy,
+    title={The NumPy Array: A Structure for Efficient Numerical Computation},
+    author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
+    journal={Computing in Science   Engineering},
+    volume={13},
+    year={2011},
+    number={2},
+    pages={22-30},
+    doi={10.1109/MCSE.2011.37},
 }
 
 @Article{matplotlib,
