### 2020-12-13 #1942 @benjaminbolling
I saw there was a missing DOI for the NumPy array, so I added this reference to paper.bib.
