+++ paper.md
@@ -22,28 +22,31 @@ bibliography: paper.bib
 
 # Summary
 
-At large research facilities and industrial complexes, there is a need of control system user interfaces. However, modern facilities also require continuous upgrading, maintenance, and development, which means that also the control systems need to be upgraded. In order to simplify the construction of control systems and diagnostics, I created the DynaGUI (Dynamic Graphical User Interface) package. The main idea of this package is to get rid of the middle-hand coding needed between hardware and the user by supplying the user with a simple GUI toolkit for generating diagnostics- and control-system GUI:s in accordance with any user’s need. Initially developed at MAX IV Laboratory [@MAXIVproj], the initial and main user for this package is controlrooms at large-scale research facilities, such as particle accelerators.
+At large research facilities and industrial complexes, there is a need for control system user interfaces. However, modern facilities also require continuous upgrading, maintenance, and development, which means that also the control systems need to be upgraded. In order to simplify the construction of control systems and diagnostics, I created the DynaGUI (Dynamic Graphical User Interface) package. The main idea of this package is to get rid of the middle-hand coding needed between hardware and the user by supplying the user with a simple GUI toolkit for generating diagnostics- and control-system GUIs in accordance with any user’s need. Initially developed at MAX IV Laboratory [@MAXIVproj], the initial and main users for this package are control rooms at large-scale research facilities, such as particle accelerators.
 
-In order to further enhance the user-friendliness of this package, a simple system for configuration files has been developed enabling users to configure it using any plain text-editor. The DynaGUI package consists of three applications:
+In order to further enhance the user-friendliness of this package, a simple system of configuration files has been developed, enabling users to configure the applications using any plain text-editor. The DynaGUI package consists of three applications:
+
 - DynaGUI TF, a true/false (boolean) dynamic control system,
+
 - DynaGUI Alarms, a dynamic diagnostics system for continuously monitoring of the values for a set of attributes, and
+
 - DynaGUI NV, a system for observing attributes' numerical-, string-, vector- or waveform values.
 
-Each DynaGUI application's layout is designed for having as high level of simplicity as possible. At the top of each application is a combobox with the list of attributes defined. Below the combobox is a button called 'Edit DynaGUI', which opens a window for configuring the DynaGUI. Below the 'Edit DynaGUI'-button is the dynamic field, in which a dynamic control panel is generated. Below the dynamic control panel field is the DynaGUI status bar, showing the last action carried out, error messages, or if an alarm is active (for DynaGUI Alarms). Below the status bar is the load and save buttons for loading or saving DynaGUI configuration files, which also have a tool-tip function showing the last loaded or saved file (in the current session). The simplest method to launch DynaGUI is via its launcher (Launcher.py), in which the user can select control system and either directly define the path to the configuration file or browse to it. Leaving it blank results in that DynaGUI will load with a predefined setup.
+Each DynaGUI application's layout is designed for simplicity. At the top of each application is a combobox with the list of attributes defined. Below the combobox is a button called 'Edit DynaGUI', which opens a window for configuring the DynaGUI. Below the 'Edit DynaGUI' button is the dynamic field, in which a dynamic control panel is generated. Below the dynamic control panel field is the DynaGUI status bar, showing the last action carried out, error messages, or if an alarm is active (for DynaGUI Alarms). Below the status bar are the load and save buttons for loading or saving DynaGUI configuration files. These also have a tool-tip function showing the last loaded or saved file (in the current session). The simplest method to launch DynaGUI is via its launcher (Launcher.py), in which the user can select a control system and either directly define the path to the configuration file or browse to it. If the field is left blank, DynaGUI will load with a predefined setup.
 
[…]
