+++ paper.bib
@@ -39,7 +39,8 @@
 @ARTICLE{numpy,  
     title={The NumPy Array: A Structure for Efficient Numerical Computation},   
     author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},  
-    journal={Computing in Science   Engineering},   
+    journal={Computing in Science   Engineering},
+    doi={10.1109/mcse.2011.37},
     volume={13},  
     year={2011},  
     number={2},  
