+++ paper.md
@@ -36,7 +36,7 @@ Each DynaGUI application's layout is designed for simplicity. At the top of each
 
 DynaGUI TF (True/False) is dynamical in the sense that the user can insert the name of any device's servers and attributes that, in current state, are True or False. The GUI then builds itself up by creating buttons for each device server that will display the boolean of the device in the combobox selected attribute. First, the application will try to connect to the device's server as entered, and then read in the attribute selected in the combobox and paint the button's background colour:  green meaning True, red meaning False, fuchsia meaning attribute not existing or not boolean, and maroon meaning that the device cannot be connected.
 
-The DynaGUI Alarms (Dynamic Alarms GUI) has been designed to monitor numerical values and notify user(s) if a condition is not fulfilled. To edit the alarms GUI, the user has to press Edit DynaGUI to open an edit-panel. In the left window of the edit-panel, the user has to define the list of devices' server domains and signals to monitor. In the right window of the edit-panel, the user has to define descriptions of the different alarm signals as they should appear in the DynaGUI Alarms control panel, as well as the sweep-time (how frequently the system should check the values). The DynaGUI Alarms' dynamic panel can be divided into three columns for each device and attribute it monitors. The first column contains the descriptions of the alarms, the second column contains the numerical values from the last sweep, the fourth column contains the alarm limit, and the third column contains a combobox with two gaps (larger than or smaller than) which points in the way it should between the read value and the alarm limit. For gap criteria that are fulfilled, the background-colour of the alarm description is painted lime-green. If a gap criteria is not fulfilled, the computer's speakers will emit an alarm, and the display will show a message in the DynaGUI Alarms' status bar. The  alarm description's background colour becomes red.
+The DynaGUI Alarms (Dynamic Alarms GUI) has been designed to monitor numerical values and notify user(s) if a condition is not fulfilled. To edit the alarms GUI, the user has to press Edit DynaGUI to open an edit-panel. In the left window of the edit-panel, the user has to define the list of devices' server domains and signals to monitor. In the right window of the edit-panel, the user has to define descriptions of the different alarm signals as they should appear in the DynaGUI Alarms control panel, as well as the sweep-time (how frequently the system should check the values). The DynaGUI Alarms' dynamic panel can be divided into three columns for each device and attribute it monitors. The first column contains the descriptions of the alarms, the second column contains the numerical values from the last sweep, the fourth column contains the alarm limit, and the third column contains a combobox with two gaps (larger than or smaller than) which points in the way it should between the read value and the alarm limit. For gap criteria that are fulfilled, the background colour of the alarm description is painted lime-green. If a gap criteria is not fulfilled, the computer's speakers will emit an alarm, and the display will show a message in the DynaGUI Alarms' status bar. The  alarm description's background colour becomes red.
 
 DynaGUI NV is the most advanced tool in the package. Each device has two columns. The first column is a button with the server address of the device, whilst the value of the selected attribute is shown in the second column, enabling users to inspect values in a simple and fast manner. DynaGUI NV also allows for launching a controller for the device (using AtkPanel for Tango Controls [@TangoCS]) or for initializing 1D plots or 2D colormaps of the selected attribute for all devices for which the attribute is valid. Each device's control panel button becomes painted in lime-green if the attribute is valid, in fuchsia if the attribute is not existing for the device and maroon if a connection to the device cannot be established. The plot initialization automatically launches for all devices for which the attribute is valid. The plotting tool has been elaborately described in the DynaGUI documentation. Features in the 1D graph tool include  plotting read values in real time, and setting up and plotting equations (or functions) by using NumPy [@numpy] of the read values as well as combining the read values with one another. Examples are shown in Figure 1. The 1D graph tool supports both scalar and vector (or waveform) plotting in real time, and plot data can be saved and loaded.
 
