### 2020-12-13 #1942 @majensen
@benjaminbolling  you'll see a PR with a few scattered edits.
The paper is missing "Figure 1", referred to in the text - not present in the repo.
Thanks
