### 2020-12-14 bc2b0a5
Invalid figure reference in paper which has been updated
