### 2020-12-15 #1942 @majensen
Ok @benjaminbolling -- I think we are on our way to the finish. The last thing to do before I call upon Those Who Must Not Be Named (the associate editors in chief) is to create an open archive of the Github repo. Zenodo.org is a good place for this - have a look at https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729 for a brief tutorial.
The main detail here is to name the archive the same as the paper ("The DynaGUI Package") and for the archive to have the same author as the paper. You'll have an archive version and a DOI -- please report those back in this thread.
