### 2020-12-26 #1942 @danielskatz
@benjaminbolling  - do you mean for the version name to be "Version 1" as opposed to "v1.0" or "1.0"?

(Also, as an FYI, the link at the end of the README to the license doesn't work, and your use of ":s" in the README is also confusing, as in "GUI:s" and "BPM:s" - if you mean these to be plurals, they should just be "GUIs" and "BPMs", I think, and if they are intended to be possessive, which I'm fairly sure you don't, "GUI's" and "BPM's" would be correct)
