### 2020-12-26 #1942 @danielskatz
In the paper, the fact that the first paragraph has some text written in the first person (I created), while the rest of the paper is written in the third paragraph (The author wants to thank) is a little confusing. Maybe change the first part to third person too?

Otherwise, this looks ready to accept.
