### 2020-12-28 a991a1b
Fixed the licence link in the README file.
