### 2020-12-28 #1942 @benjaminbolling
Dear @danielskatz, thank you for your comments. I have change the use of ":s" to just "s" (when it was used to describe objects in plural). Also, the first paragraph in the user guide is rewritten to third paragraph and the acknowledgements has been updated to sound less confusing.

In accordance with this, a new version of the files have been archived on Zenodo with a new DOI.
