### 2020-12-28 #1942 @benjaminbolling
@majensen : Please update Zenodo archive to:

Archive version: v1.0
DOI: 10.5281/zenodo.4396035

Thank you!
