### 2020-12-28 #1942 @danielskatz
@benjaminbolling  - I don't see any changes in the paper - the last update seems to be 14 days ago

https://github.com/benjaminbolling/DynaGUI/blob/master/paper.md
