### 2020-12-28 f8e5b9b
paper.md : removed I and fixed acknowledgements
