### 2020-12-28 #1942 @benjaminbolling
Removed .git from zip file and updated paper.md in the archived file.

Please update Zenodo archive to:

Archive version: v1.0.1
DOI: 10.5281/zenodo.4396227
