### 2020-12-28 #1942 @danielskatz
Thanks - there's no need to update the repository for a change in the paper, but it's also fine to
